% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed-class.R
\docType{class}
\name{bed-class}
\alias{bed-class}
\alias{bed_RC}
\alias{bed}
\title{Class bed}
\usage{
bed(bedfile)
}
\arguments{
\item{bedfile}{Path to file with extension ".bed" to read.
You need the corresponding ".bim" and ".fam" in the same directory.}
}
\description{
A reference class for storing a pointer to a mapped version of a bed file.
}
\details{
A \code{bed} object has many field:
\itemize{
\item \code{$address}: address of the external pointer containing the underlying
C++ object, to be used as a \code{XPtr<bed>} in C++ code
\item \code{$extptr}: use \code{$address} instead
\item \code{$bedfile}: path to the bed file
\item \code{$bimfile}: path to the corresponding bim file
\item \code{$famfile}: path to the corresponding fam file
\item \code{$nrow}: number of samples in the bed file
\item \code{$ncol}: number of variants in the bed file
}
}

\examples{
bedfile <- system.file("extdata", "example-missing.bed", package = "bigsnpr")
obj.bed <- bed(bedfile)

}
