% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed-mat-acc.R
\name{[,bed,ANY,ANY,ANY-method}
\alias{[,bed,ANY,ANY,ANY-method}
\alias{[,bed_light,ANY,ANY,ANY-method}
\title{Accessor methods for class \code{bed}.}
\usage{
\S4method{[}{bed,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{bed_light,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{A \link[=bed-class]{bed} object.}

\item{i}{A vector of indices (or nothing). You can use positive and negative
indices, and also logical indices (that are recycled).}

\item{j}{A vector of indices (or nothing). You can use positive and negative
indices, and also logical indices (that are recycled).}

\item{...}{Not used. Just to make \link{nargs} work.}

\item{drop}{Whether to drop dimensions (to a vector) when a dimension is 1.
Default is \code{TRUE}.}
}
\description{
Accessor methods for class \code{bed}.
}
\examples{
bedfile <- system.file("extdata", "example-missing.bed", package = "bigsnpr")
(obj.bed <- bed(bedfile))
obj.bed[1:5, 1]
obj.bed[1:5, 1:2]
typeof(obj.bed[1, 1])

}
\keyword{internal}
