% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed-class.R
\name{bed-methods}
\alias{bed-methods}
\alias{dim,bed-method}
\alias{length,bed-method}
\title{Methods for the bed class}
\usage{
\S4method{dim}{bed}(x)

\S4method{length}{bed}(x)
}
\arguments{
\item{x}{Object of type \code{bed}.}
}
\description{
Methods for the bed class

Dimension methods for class \code{bed}.
Methods \code{nrow()} and \code{ncol()} are automatically defined with \code{dim()}.
}
