% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed-mult-vec.R
\name{big_prodVec.bed}
\alias{big_prodVec.bed}
\title{Product with a vector}
\usage{
\method{big_prodVec}{bed}(X, y.col, ind.row = rows_along(X),
  ind.col = cols_along(X), center = NULL, scale = NULL)
}
\arguments{
\item{X}{A \link{bed} object.}

\item{y.col}{A vector of same size as \code{ind.col}.}

\item{ind.row}{An optional vector of the row indices that are used.
If not specified, all rows are used. \strong{Don't use negative indices.}}

\item{ind.col}{An optional vector of the column indices that are used.
If not specified, all columns are used. \strong{Don't use negative indices.}}

\item{center}{Vector of same length of \code{ind.col} to subtract from columns
of \code{X}.}

\item{scale}{Vector of same length of \code{ind.col} to divide from columns
of \code{X}.}
}
\value{
\eqn{X \cdot y}.
}
\description{
Product between a "bed" object and a vector.

Missing values are replaced by 0 (after centering), as if they
had been imputed using parameter \code{center}.
}
\examples{
bedfile <- system.file("extdata", "example-missing.bed", package = "bigsnpr")
obj.bed <- bed(bedfile)

y.col <- rep(1, ncol(obj.bed))
big_prodVec(obj.bed, y.col)

}
