% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigSNP.R
\name{bigSNP-class}
\alias{bigSNP-class}
\alias{bigSNP}
\title{S3 class bigSNP}
\description{
A class for representing infos on massive SNP arrays.
}
\details{
A named list with at least 5 slots: \describe{
\item{genotypes}{A filebacked \link[=big.matrix-class]{big.matrix}
of type \code{char} (one byte signed integer), representing genotypes
(read from a ".bed" file). Each element is either \code{0}, \code{1}, \code{2} or \code{NA}.
Rows are individuals and columns are SNPs.}
\item{fam}{A \code{data.frame} containing some information on the SNPs
(read from a ".fam" file).}
\item{map}{A \code{data.frame} giving some information on the individuals
(read from a ".bim" file).}
\item{backingfile}{The root name for the backing files of the object.}
\item{backingpath}{The path to the directory containing the backing files.}
}
}
\seealso{
\link{snp_readBed}
}
\keyword{class}
