% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigsnpr.R
\docType{package}
\name{bigsnpr-package}
\alias{bigsnpr}
\alias{bigsnpr-package}
\title{bigsnpr: Predictive Analysis of Massive SNP Arrays}
\arguments{
\item{G}{A \link[=BM.code.descriptor-class]{BM.code.descriptor}.
Typically the slot \code{genotypes} of a \code{bigSNP}.
You shouldn't have missing values in your data.}

\item{x}{A \link[=bigSNP-class]{bigSNP}.}

\item{has.pheno}{Does \strong{\code{x}} has well-specified phenotypes?
i.e. \code{x$fam$affection} is a vector with exactly two values
(and the value coding for cases is larger than the value coding
for controls). So, in particular, no missing value is allowed.
Default is \code{TRUE}.}

\item{nploidy}{Number of trials, parameter of the binomial distribution.
Default is \code{2}, which corresponds to diploidy, such as for the human genome.}

\item{ind.row}{An optional vector of the row (individuals) indices that
are used. If not specified, all rows are used.}

\item{ind.col}{An optional vector of the column (SNP) indices that are used.
If not specified, all columns are used.}

\item{block.size}{Maximum number of loci read at once (for all individuals).}

\item{ncores}{Number of cores used. Default doesn't use parallelism.}
}
\description{
Easy-to-use, efficient, flexible and scalable tools
for predictive analysis based on SNP arrays.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://privefl.github.io/bigsnpr}
  \item Report bugs at \url{https://github.com/privefl/bigsnpr/issues}
}

}
\author{
\strong{Maintainer}: Florian Privé \email{florian.prive.21@gmail.com}

Other contributors:
\itemize{
  \item Michael Blum \email{michael.blum@univ-grenoble-alpes.fr} [thesis advisor]
  \item Hugues Aschard \email{hugues.aschard@pasteur.fr} [thesis advisor]
}

}
