% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{counts}
\alias{CountByPheno}
\alias{CountNAByRow}
\alias{counts}
\title{Counts.}
\usage{
CountByPheno(x)

CountNAByRow(x)
}
\arguments{
\item{x}{A \code{bigSNP}.}
}
\value{
\code{CountByPheno}: An integer matrix
of size 8*m (m is the number of SNPs).

\code{CountNAByRow}: An integer vector
of size n (n is the number of individuals).
}
\description{
\code{CountByPheno}: counts the number of 0, 1, 2 and NA
by SNP and phenotype (cases/controls).

\code{CountNAByRow}: counts the number of NA
by individuals.
}
\examples{
# constructing a fake genotype big.matrix
a <- big.matrix(10, 15, type = "char", shared = FALSE)
a[] <- sample(c(0, 1, 2, NA), 150, TRUE)
print(a[,])

# constructing a fake incomplete bigSNP with 10 individuals and 15 SNPs
# where the 5 first individuals are cases and the 5 last are controls.
fake <- list()
class(fake) <- "bigSNP"
fake$genotypes <- a
fake$fam$pheno <- c(rep(1, 5), rep(-1, 5))

# Get counts
print(CountByPheno(fake))
print(CountNAByRow(fake))
}

