% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval.R
\name{crossval.ext}
\alias{crossval.ext}
\title{K-fold cross-validation.}
\usage{
crossval.ext(x, fun.model, K)
}
\arguments{
\item{x}{A \link[=bigSNP-class]{bigSNP}.}

\item{fun.model}{Function that is used to train and
predict a particular model. Its first 3 parameters
should be \code{x}, \code{ind.train} and \code{ind.test} and it should
have a parameter \code{fun.model} with \code{NULL} as default value.}

\item{K}{Number of turns in the cross-validation.}
}
\value{
A list of each turn's score.
}
\description{
K-fold cross-validation of \code{fun.model} on \code{x}.
}
