% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval.R
\name{crossval.int}
\alias{crossval.int}
\title{K-fold cross-validation.}
\usage{
crossval.int(x, fun.model, ind.train, K)
}
\arguments{
\item{x}{A \link[=bigSNP-class]{bigSNP}.}

\item{fun.model}{Function that is used to train and
predict a particular model. Its first 3 parameters
should be \code{x}, \code{ind.train} and \code{ind.test} and it should
have a parameter \code{fun.model} with \code{NULL} as default value.}

\item{ind.train}{Indices used as training set by the model.}

\item{K}{Number of turns in the cross-validation.}
}
\value{
The position of the parameter which has given
the best scores (in average) of the model.
}
\description{
Internal K-fold cross-validation of \code{fun.model} on \code{x}
in order to tune hyper-parameters.
}

