% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectPCA-bed.R
\name{download_1000G}
\alias{download_1000G}
\title{Download 1000G}
\usage{
download_1000G(dir, overwrite = FALSE)
}
\arguments{
\item{dir}{The directory where to put the downloaded files.}

\item{overwrite}{Whether to overwrite files when downloading and unzipping?
Default is \code{FALSE}, so this function does not do anything the second time.}
}
\value{
The path of the downloaded bed file.
}
\description{
Download 1000 genomes project (phase 3) data in PLINK bed/bim/fam format,
including 2493 (mostly unrelated) individuals
and ~1.4M SNPs in common with HapMap3.
}
