% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-software.R
\name{download_beagle}
\alias{download_beagle}
\title{Download Beagle 4.1}
\usage{
download_beagle(dir = tempdir())
}
\arguments{
\item{dir}{The directory where to put the Beagle Java Archive.
Default is a temporary directory.}
}
\value{
The path of the downloaded Beagle Java Archive.
}
\description{
Download Beagle 4.1 from
\url{https://faculty.washington.edu/browning/beagle/beagle.html}
}
