% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-sumstats.R
\name{eigen_halfinv}
\alias{eigen_halfinv}
\title{Eigen decomposition to get x^-0.5}
\usage{
eigen_halfinv(x, prop_eigs, tol = 1e-04)
}
\arguments{
\item{x}{A square symmetric matrix.}

\item{prop_eigs}{The proportion of eigen components to keep.}

\item{tol}{The minimum eigenvalue allowed; others are discarded.}
}
\value{
An approximation of x^-0.5.
}
\description{
Eigen decomposition to get x^-0.5
}
\examples{
x <- tcrossprod(matrix(1:4, 2))
eig_scaled <- bigsnpr:::eigen_halfinv(x, prop_eigs = 1)
all.equal(solve(x), tcrossprod(eig_scaled))

}
