% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcimpute.R
\name{impute-qc-sub}
\alias{Impute}
\alias{QC}
\alias{impute-qc-sub}
\alias{sub.bigSNP}
\title{Imputation, quality control and subset for a "bigSNP".}
\usage{
sub.bigSNP(x, ind.row = seq(nrow(x$genotypes)),
  ind.col = seq(ncol(x$genotypes)))

QC(x, row.cr.min = 0.95, col.cr.min = 0.95, hwe.pval = 1e-06,
  maf.min = NULL, rm.sex = FALSE)

Impute(x, ncores = 1, verbose = FALSE)
}
\arguments{
\item{x}{A \code{bigSNP}.}

\item{ind.row}{Indices of the rows (individuals) to keep.
Negative indices can be used to exclude row indices.
Default: keep them all.}

\item{ind.col}{Indices of the columns (SNPs) to keep.
Negative indices can be used to exclude column indices.
Default: keep them all.}

\item{row.cr.min}{Minimun individuals' call rate that is allowed.
Default is 95\%.}

\item{col.cr.min}{Minimum SNPs' call rate that is allowed. Default is 95\%.}

\item{hwe.pval}{Level threshold (allowed type-I error) to test deviations
from Hardy–Weinberg equilibrium (HWE) from controls only. Default is \code{1e-6}.}

\item{maf.min}{Minimum Minor Allele Frequency that is allowed.
Usually, \code{0.01} is used. Default only removes SNPs that have a zero MAF.}

\item{rm.sex}{Keep only SNPs on the first 22 chrosmosomes? Default is \code{FALSE}.}

\item{ncores}{Number or cores used.
Default doesn't use parallelism.}

\item{verbose}{Print progress? Default is \code{FALSE}.}
}
\value{
A new \code{bigSNP}.
}
\description{
\code{sub.bigSNP}: a function
to get a subset of an object of class \code{bigSNP}.

\code{QC}: Quality control (filters)
for a \code{bigSNP} resulting in a \code{bigSNP} of lower dimension.

\code{Impute}: Imputation function
for a \code{bigSNP}.
}
\examples{
\dontrun{

  bedfile <- system.file("extdata", "example.bed", package = "bigsnpr")

  # Creating directory for backing files
  if (!dir.exists("backingfiles")) dir.create("backingfiles")

  if (!file.exists("backingfiles/test_doc.bk")) {
    # Reading the bedfile and storing the data
    test <- BedToBig(bedfile, 50, "test_doc")
  } else {
    # Loading it from backing files
    test <- AttachBigSNP("test_doc")
  }

  # dimensions
  print(rbind(dim(test$genotypes), dim(test$fam), dim(test$map)))

  # taking only the first 50 individuals and 500 SNPs at random
  ind.row <- 1:50
  ind.col <- sort(sample(ncol(test$genotypes), 500))
  test2 <- sub.bigSNP(test, ind.row, ind.col)

  # dimensions
  print(rbind(dim(test2$genotypes), dim(test2$fam), dim(test2$map)))
  # new backing files
  print(test2$backingfile)
  print(test2$backingpath)


  # removing the 100th and 120th individuals
  ind.row.del <- c(100, 120)
  test3 <- sub.bigSNP(test, -ind.row.del)

  # dimensions
  print(rbind(dim(test3$genotypes), dim(test3$fam), dim(test3$map)))
  print(test3$fam$sample.ID[98:120])
  # new backing files
  print(test3$backingfile) # incrementation of the file's number
}
}
\seealso{
\link[=bigSNP-class]{bigSNP}
}

