% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clumping.R
\docType{data}
\name{LD.wiki34}
\alias{LD.wiki34}
\title{Long-range LD regions}
\format{
A data frame with 34 rows (regions) and 4 variables:
\itemize{
\item \code{Chr}: region's chromosome
\item \code{Start}: starting position of the region (in bp)
\item \code{Stop}: stopping position of the region (in bp)
\item \code{ID}: some ID of the region.
}
}
\usage{
LD.wiki34
}
\description{
34 long-range Linkage Disequilibrium (LD) regions for the human genome
based on some \href{https://goo.gl/0Ou7uI}{wiki table}.
}
\keyword{datasets}
