% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAX3.R
\name{MAX3}
\alias{MAX3}
\title{MAX3 statistic.}
\usage{
MAX3(x, ind.train = seq(nrow(x$genotypes)))
}
\arguments{
\item{x}{A \code{bigSNP}.}

\item{ind.train}{An optional vector of the row indices that are used,
for the training part.
If not specified, all data are used.}
}
\value{
A numeric vector of the column statistics.
}
\description{
Counts the number of 0, 1, 2
by SNP and phenotype (cases/controls) and then
compute the MAX3 statistic.
}
\examples{
# constructing a fake genotype big.matrix
a <- big.matrix(10, 15, type = "char", shared = FALSE)
a[] <- sample(c(0, 1, 2), 150, TRUE)
print(a[,])

# constructing a fake incomplete bigSNP with 10 individuals and 15 SNPs
# where the 5 first individuals are cases and the 5 last are controls.
fake <- list()
class(fake) <- "bigSNP"
fake$genotypes <- a
fake$fam$pheno <- c(rep(1, 5), rep(-1, 5))

# Get MAX3 statistics
print(MAX3(fake))
}

