% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval.R
\name{model.fun}
\alias{model.fun}
\title{Create a function with specified parameters.}
\usage{
model.fun(fun, ...)
}
\arguments{
\item{fun}{The function you want to use.}

\item{...}{Replace this by a list of parameters of
\code{fun} you want to set.}
}
\value{
A new function, derived from \code{fun}.
}
\description{
Create a function with specified parameters.
}
\examples{
round3 <- model.fun(round, digits = 3)
round3(4.5588877)
}
\seealso{
\link{crossval.ext} \link{crossval.int}
}
