% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{modif-save}
\alias{GetPhenos}
\alias{GetPops}
\alias{SaveModifs}
\alias{modif-save}
\title{Modify a "bigSNP" and save these modifications.}
\usage{
SaveModifs(x)

GetPops(x, pop.files, col.sample.ID, col.family.ID)

GetPhenos(x, coded01 = FALSE)
}
\arguments{
\item{x}{A \code{bigSNP}.}

\item{pop.files}{Character vector of file names where to
find the population of the individuals of the study.}

\item{col.sample.ID}{Index of the column containing the
sample IDs to match with those of the study.}

\item{col.family.ID}{Index of the column containt the populations.}

\item{coded01}{Are the phenotypes coded 0 (unaffected) / 1 (affected)
rather than respectively 1 and 2? Default is \code{FALSE}.}
}
\value{
The (modified) \code{bigSNP}.
}
\description{
Modify a \code{bigSNP} and save these modifications
in the corresponding .rds file. As \code{bigSNP} is an S3 class, you
can add any slot you want to this class, then use \code{SaveModifs} to
save modifications.

\code{GetPops}: Get populations of individuals by matching
from an external file if it wasn't specified in PLINK files.
It will modify the slot "fam$family.ID".

\code{GetPhenos}: Modify phenotypes to be -1 (unaffected),
1 (affected) or NA (missing).
It will add an slot called "pheno" to the slot "fam".
}

