% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRS.R
\name{PRS1}
\alias{PRS1}
\title{Title}
\usage{
PRS1(x, ind.train, ind.test, fun.stats, weighted = FALSE, thr.list = 0,
  pruning = FALSE, K = 10, fun.model = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
