% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning.R
\name{Prune}
\alias{Prune}
\title{LD pruning for a "bigSNP"}
\usage{
Prune(x, ind.train = seq(nrow(X)), fun.stats, thr.pvalue = 1, size = 2000,
  thr.corr = 0.2, ncores = 1)
}
\arguments{
\item{x}{A \link[=bigSNP-class]{bigSNP}.}

\item{ind.train}{An optional vector of the row indices that are used,
for the training part. If not specified, all data are used.}

\item{thr.pvalue}{Threshold on \eqn{-log_{10}(p-value)} to assess
which SNPs are kept. Here, it has the purpose to accelerate computations.
Default is 1.}

\item{size}{Radius of the window's size for the LD evaluations.}

\item{thr.corr}{Threshold on the correlation between two SNPs.
SNPs which are too correlated with another SNP which is more correlated
with the disease are pruned.}

\item{ncores}{Number or cores used. Default doesn't use parallelism.}

\item{S}{Numeric vector of summary statistics computed
only with \code{ind.train}.}

\item{pS}{Numeric vector of p-values associated with S.
\strong{\code{pS} needs to be computed through a decreasing function of \code{S}}.
For example, if \code{S} follows the standard normal distribution,
you should use \code{abs(S)} instead and compute
\code{pS = 2 * pnorm(abs(S), lower.tail = FALSE)}.}
}
\description{
LD pruning for a \code{bigSNP}.
}
\examples{
# Creating directory for backing files
if (!dir.exists("backingfiles")) dir.create("backingfiles")

if (!file.exists("backingfiles/test_doc.bk")) {
  # Reading the bedfile and storing the data
  bedfile <- system.file("extdata", "example.bed", package = "bigsnpr")
  test <- BedToBig(bedfile, 50, "test_doc")
} else {
  # Loading it from backing files
  test <- AttachBigSNP("test_doc")
}

# generating random phenotypes
test$fam$pheno <- sample(c(-1, 1), size = nrow(test$fam),
                         replace = TRUE)

ind <- Prune(test)
# ind <- Prune(test, ncores = 2)

R2 <- RsqClass(test$genotypes, test$fam$pheno)

par.save <- par(mfrow = c(1, 2))

plot(R2, type = 'h')
plot(ind, R2[ind], type = "h")

par(par.save)
}

