% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning.R
\name{pruning-clumping}
\alias{pruning-clumping}
\alias{snp_clumping}
\alias{snp_pruning}
\alias{snp_indLRLDR}
\title{LD pruning and clumping}
\usage{
snp_clumping(G, infos.chr, ind.row = rows_along(G), S = NULL, size = 1000,
  is.size.in.kb = FALSE, infos.pos = NULL, thr.r2 = 0.5, exclude = NULL,
  ncores = 1)

snp_pruning(G, infos.chr, ind.row = rows_along(G), size = 50,
  is.size.in.kb = FALSE, infos.pos = NULL, thr.r2 = 0.5, exclude = NULL,
  nploidy = getOption("bigsnpr.nploidy"), ncores = 1)

snp_indLRLDR(infos.chr, infos.pos, LD.regions = LD.wiki34)
}
\arguments{
\item{G}{A \link[=BM.code.descriptor-class]{BM.code.descriptor}.
Typically the slot \code{genotypes} of a \code{bigSNP}.
You shouldn't have missing values in your data.}

\item{ind.row}{An optional vector of the row (individuals) indices that
are used. If not specified, all rows are used.}

\item{S}{A vector of column statistics which express the importance
of each SNP (the more important is the SNP, the greater should be
the corresponding statistic). For example, if \code{S} follows the standard normal
distribution, and significant means significantly different from 0,
you should probably use \code{abs(S)} instead. If not specified, the MAF
is computed and used.}

\item{size}{This parameter should be adjusted with respect to the number of SNPs.
\itemize{
\item for clumping: \strong{Radius} of the window's size for the LD evaluations.
Default is \code{1000} (I use this for a chip of 500K SNPs).
\item for pruning: \strong{Diameter} of the window's size for the LD evaluations.
Default is \code{50} (as in PLINK 1.07).
}}

\item{thr.r2}{Threshold over the squared correlation between two SNPs.
Default is \code{0.5}.}

\item{exclude}{Vector of indices of SNPs to exclude anyway. For example,
can be used to exclude long-range LD regions (see Price2008). Another use
can be for thresholding with respect to p-values associated with \code{S}.}

\item{ncores}{Number of cores used. Default doesn't use parallelism.}

\item{nploidy}{Number of trials, parameter of the binomial distribution.
Default is \code{2}, which corresponds to diploidy, such as for the human genome.}

\item{LD.regions}{A \code{data.frame} with columns "Chr", "Start" and "Stop".
Default use the table of 34 long-range LD regions that you can find
\href{https://goo.gl/0Ou7uI}{there}.}
}
\value{
\itemize{
\item \code{snp_pruning} & \code{snp_pruning}: SNP indices which are \strong{kept}.
\item \code{snp_indLRLDR}: SNP indices to be used as (part of) the
\strong{\code{exclude}} parameter of \code{snp_pruning} or \code{snp_clumping}.
}
}
\description{
For a \code{bigSNP}:
\itemize{
\item \code{snp_pruning}: LD pruning. Similar to "\code{--indep-pairwise size 1 thr.r2}" in
\href{http://pngu.mgh.harvard.edu/~purcell/plink/summary.shtml#prune}{PLINK 1.07}
(\code{step} is fixed to 1).
\item \code{snp_clumping}: LD clumping.
\item \code{snp_indLRLDR}: Get SNP indices of long-range LD regions for the
human genome.
}
}
\details{
I recommend to use clumping rather than pruning. See
\href{https://privefl.github.io/bigsnpr/articles/pruning-vs-clumping.html}{this article}.
}
\examples{
set.seed(1)

test <- snp_readExample() # need another example
X <- test$genotypes
print(n <- nrow(X))
print(m <- ncol(X))

# pruning / clumping with MAF
ind.keep <- snp_pruning(test, thr.corr = 0.1)
print(length(ind.keep) / m)

p <- big_colstats(X)$sum / (2 * n)
maf <- pmin(p, 1 - p)
ind.keep2 <- snp_clumping(test, S = maf, thr.corr = 0.1)

# clumping for PRS
test$fam$affection <- sample(1:2, size = n, replace = TRUE)
zcatt <- snp_MAX3(test, val = 0.5)
R2 <- big_univRegLin(X, test$fam$affection)["R2", ]
plot(zcatt$S, n * R2) # same thing
ind.keep3 <- snp_clumping(test, S = zcatt$S, thr.corr = 0.1)
print(length(ind.keep3))
# thresholding though the exclude parameter
ind.keep4 <- snp_clumping(test, S = zcatt$S, thr.corr = 0.1,
                          exclude = which(-log10(zcatt$pS) > 1))
print(length(ind.keep4))
}
\references{
Price AL, Weale ME, Patterson N, et al.
Long-Range LD Can Confound Genome Scans in Admixed Populations.
Am J Hum Genet. 2008;83(1):132-135.
\url{http://dx.doi.org/10.1016/j.ajhg.2008.06.005}
}
