% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning.R
\name{pruning}
\alias{pruning}
\title{LD clumping}
\arguments{
\item{fun.stats}{A function that takes a \code{big.matrix} \strong{\code{X}} and
\strong{\code{ind.train}} as parameters and returns a named list of \strong{\code{S}} and
\strong{\code{pS}} for every column, which are statistics and associated p-values.
\strong{\strong{\code{pS}} needs to be computed through a decreasing function of \code{S}}.
For example, if \strong{\code{S}} follows the standard normal distribution,
you should use \code{abs(S)} instead and compute
\code{pS = 2 * pnorm(abs(S), lower.tail = FALSE)}.}

\item{size}{Radius of the window's size for the LD evaluations.
Default is \code{1000}, which seems pretty conservative for a standard
chip of less than a million SNPs.}

\item{thr.pvalue}{Threshold on \eqn{-log_{10}(p-value)} to assess
which SNPs are kept. A default of \code{1} is very conservative and
has the purpose to accelerate computations.}

\item{thr.corr}{Threshold on the correlation between two SNPs.
SNPs which are too correlated with another SNP which is more correlated
with the disease are pruned.}

\item{exclude}{Vector of indices of SNPs to exclude anyway. For example,
can be used to exclude long-range LD regions (see Price2008)}
}
\description{
LD clumping (and thresholding) for a \code{bigSNP}.
}
\examples{
# Creating directory for backing files
if (!dir.exists("backingfiles")) dir.create("backingfiles")

if (!file.exists("backingfiles/test_doc.bk")) {
  # Reading the bedfile and storing the data
  bedfile <- system.file("extdata", "example.bed", package = "bigsnpr")
  test <- BedToBig(bedfile, 50, "test_doc")
} else {
  # Loading it from backing files
  test <- AttachBigSNP("test_doc")
}

# generating random phenotypes
test$fam$pheno <- sample(c(-1, 1), size = nrow(test$fam),
                         replace = TRUE)

ind <- Prune(test)
# ind <- Prune(test, ncores = 2)

R2 <- RsqClass(test$genotypes, test$fam$pheno)

par.save <- par(mfrow = c(1, 2))

plot(R2, type = 'h')
plot(ind, R2[ind], type = "h")

par(par.save)
}
\references{
Price AL, Weale ME, Patterson N, et al.
Long-Range LD Can Confound Genome Scans in Admixed Populations.
Am J Hum Genet. 2008;83(1):132-135.
\link{http://dx.doi.org/10.1016/j.ajhg.2008.06.005}.
}

