% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcimpute.R
\name{qcimpute}
\alias{Impute}
\alias{QC}
\alias{qcimpute}
\alias{sub.bigSNP}
\title{Quality control, subset and imputation for a "bigSNP".}
\usage{
QC(x)

sub.bigSNP(x, ind.row = NULL, ind.col = NULL)

Impute(x, ncores = 1)
}
\arguments{
\item{x}{A \code{bigSNP}.}

\item{ind.row}{Indices of the rows (individuals) to keep.
Negative indices can be used to exclude row indices.
Default: keep them all.}

\item{ind.col}{Indices of the columns (SNPs) to keep.
Negative indices can be used to exclude column indices.
Default: keep them all.}

\item{ncores}{Number or cores used.
Default doesn't use parallelism.}
}
\value{
A new \code{bigSNP}.
}
\description{
\code{QC}: Quality control (filters)
for a \code{bigSNP} resulting
in a \code{bigSNP} of lower dimension.

\code{sub.bigSNP}: a function
to get a subset of an object of class \code{bigSNP}.

\code{Impute}: Imputation function
for a \code{bigSNP}.
}
\examples{
#TODO
}
\seealso{
\code{\link{bigSNP}}
}

