% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqman.R
\name{qq2}
\alias{qq2}
\title{Q-Q plot}
\usage{
qq2(pvalues, ...)
}
\arguments{
\item{pvalues}{A numeric vector of p-values or a data.frame (or just a list)
with a named vector called "p.value".}

\item{...}{Other arguments passed to plot().}
}
\value{
\code{NULL}. Creates a Q-Q plot.
}
\description{
Creates a quantile-quantile plot from p-values from a GWAS study based on
the package \strong{qqman}.
}
