% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reglin.R
\name{reglin}
\alias{BetasClass}
\alias{BetasReg}
\alias{RsqClass}
\alias{RsqReg}
\alias{reglin}
\title{Operations of linear regression}
\usage{
RsqReg(X, y, ind.train = seq(nrow(X)))

RsqClass(X, y, ind.train = seq(nrow(X)))

BetasReg(X, y, ind.train = seq(nrow(X)))

BetasClass(X, y, ind.train = seq(nrow(X)))
}
\arguments{
\item{X}{A \code{\link[bigmemory]{big.matrix}}.}

\item{y}{Either \itemize{
\item a vector of \{-1, 1\} in the case of classification (suffix Class),
\item a vector of more than two unique values in the case of regression (suffix Reg).
}}

\item{ind.train}{A vector of the row indices that are used.}
}
\value{
\itemize{
\item \code{RsqReg} and \code{RsqClass} return a vector of
the coefficient of determination \eqn{R^2} of each regression,
\item \code{BetasReg} and \code{BetasClass} return a vector of
the slopes \eqn{\beta_1} of each regression.
}
}
\description{
Operations of \bold{univariate} linear regression on all the columns of a big.matrix.\cr
In the case of classification, weights are used to make the results independent
from the proportion of cases / controls (see parameter \emph{weights} of \code{\link[stats]{lm}}).
}

