% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom-scaling.R
\name{scale_binom}
\alias{scale_binom}
\title{Binomial(2, p) scaling}
\usage{
scale_binom(X, ind.train)
}
\arguments{
\item{X}{A \link[bigmemory:big.matrix-class]{big.matrix}.
You shouldn't have missing values in your data.}

\item{ind.train}{An optional vector of the row indices that are used,
for the training part. If not specified, all data are used.}
}
\value{
A named list of two vectors \strong{\code{mean}} and \strong{\code{sd}}
which are as long as the number of columns of \strong{\code{X}}.
}
\description{
Binomial(2, p) scaling
}
\details{
You will probably not use this function as is
but as the \strong{\code{fun.scaling}} parameter of other functions
of package \strong{\code{bigstatsr}}.
}
\examples{
a <- matrix(0, 43, 17)
p <- 0.2
a[] <- rbinom(length(a), 2, p)
X <- as.big.matrix(a, type = "char")
X.svd <- bigstatsr::big_SVD(X, fun.scaling = scale_binom)
str(X.svd)
plot(X.svd$means)
abline(h = 2 * p, col = "red")
plot(X.svd$sds)
abline(h = sqrt(2 * p * (1 - p)), col = "red")
}
\references{
This scaling is widely used for SNP arrays.
Patterson N, Price AL, Reich D (2006)
Population Structure and Eigenanalysis.
PLoS Genet 2(12): e190.
\link{http://dx.doi.org/10.1371/journal.pgen.0020190}.
}

