% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match-alleles.R
\name{snp_ancestry}
\alias{snp_ancestry}
\title{Estimation of ancestry proportions}
\usage{
snp_ancestry(freq, info_freq_ref)
}
\arguments{
\item{freq}{Vector of frequencies from which to estimate ancestry proportions.}

\item{info_freq_ref}{A data frame (or matrix) with the set of frequencies to
be used as reference (one population per column).}
}
\value{
A list of four elements:
\itemize{
\item \verb{$coef}: vector of coefficients representing the ancestry proportions,
\item \verb{$pred}: vector of predicted frequencies,
\item \verb{$cor_pred}: correlation between optimal solution \verb{$pred} and \code{freq},
\item \verb{$all_cor}: all correlations between \code{freq} and \code{info_freq_ref}.
}
}
\description{
Estimation of ancestry proportions. Make sure to match summary statistics
using \code{\link[=snp_match]{snp_match()}} (and reversing frequencies when needed).
}
\examples{
X <- matrix(rbeta(3000, 1, 3), ncol = 3)
y <- X \%*\% c(0.3, 0.1, 0.6)
res <- snp_ancestry_prop(y, X)
str(res)

}
