% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readplink.R
\name{snp_attachExtdata}
\alias{snp_attachExtdata}
\title{Attach a "bigSNP" for examples and tests}
\usage{
snp_attachExtdata()
}
\value{
The example "bigSNP", filebacked in the "/tmp/" directory.
}
\description{
Attach a "bigSNP" for examples and tests
}
