% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clumpedSVD.R
\name{snp_clumpedSVD}
\alias{snp_clumpedSVD}
\title{Title}
\usage{
snp_clumpedSVD(G, ind.row = rows_along(G), ind.col = cols_along(G),
  fun.scaling = snp_scaleBinom(), thr.r2.init = 0.2, thr.r2.step = 2,
  size = 500, k = 10, roll.size = 50, int.min.size = 20, ncores = 1)
}
\arguments{
\item{ncores}{}
}
\description{
Title
}
