% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning.R
\name{snp_clumping}
\alias{snp_clumping}
\title{LD clumping}
\usage{
snp_clumping(x, ind.train = seq(nrow(X)), fun.stats, thr.pvalue = 0,
  size = 500, thr.corr = 0.2, exclude = NULL, ncores = 1)
}
\arguments{
\item{x}{A \link[=bigSNP-class]{bigSNP}.}

\item{ind.train}{An optional vector of the row indices that are used,
for the training part. If not specified, all data are used.}

\item{fun.stats}{A function that takes a \code{big.matrix} \strong{\code{X}} and
\strong{\code{ind.train}} as parameters and returns a named list of \strong{\code{S}} and
\strong{\code{pS}} for every column, which are statistics and associated p-values.
\strong{\strong{\code{pS}} needs to be computed through a decreasing function of \code{S}}.
For example, if \strong{\code{S}} follows the standard normal distribution,
you should use \code{abs(S)} instead and compute
\code{pS = 2 * pnorm(abs(S), lower.tail = FALSE)}.}

\item{thr.pvalue}{Threshold on \eqn{-log_{10}(p-value)} to assess
which SNPs are kept. A value of \code{1} is very conservative and
has the purpose to accelerate computations.
A value of \code{0} means no thresholding and is the default.}

\item{size}{Radius of the window's size for the LD evaluations.
Default is \code{500} for clumping and \code{50} for pruning. This should be
adjusted for different number of SNPs (this corresponds to the defaults
I use for a chip of 500K SNPs).}

\item{thr.corr}{Threshold on the correlation between two SNPs.
Default is \code{0.2}.}

\item{exclude}{Vector of indices of SNPs to exclude anyway. For example,
can be used to exclude long-range LD regions (see Price2008).}

\item{ncores}{Number or cores used. Default doesn't use parallelism.}
}
\description{
LD clumping (and thresholding) for a \code{bigSNP}.
}
\details{
I recommend to use clumping rather than pruning.
See \url{https://privefl.github.io/bigsnpr/articles/pruning-vs-clumping.html}.
}
\references{
Price AL, Weale ME, Patterson N, et al.
Long-Range LD Can Confound Genome Scans in Admixed Populations.
Am J Hum Genet. 2008;83(1):132-135.
\url{http://dx.doi.org/10.1016/j.ajhg.2008.06.005}.
}

