% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigSNP-class.R, R/impute.R
\docType{data}
\name{CODE_012}
\alias{CODE_012}
\alias{CODE_DOSAGE}
\alias{CODE_IMPUTE_PRED}
\title{CODE_012: code genotype calls (3) and missing values.}
\format{
An object of class \code{numeric} of length 256.

An object of class \code{numeric} of length 256.

An object of class \code{numeric} of length 256.
}
\usage{
CODE_012

CODE_DOSAGE

CODE_IMPUTE_PRED
}
\description{
CODE_012: code genotype calls (3) and missing values.

CODE_DOSAGE: code genotype calls and missing values (4), and imputed calls (3)
and imputed allele dosages rounded to two decimal places (201).

CODE_IMPUTE_PRED: code genotype calls and missing values (4),
and imputed calls (3).
}
\keyword{datasets}
