% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R, R/corr_with_cov.R
\name{snp_cor}
\alias{snp_cor}
\alias{bed_cor}
\alias{snp_cor_with_cov}
\title{Correlation matrix}
\usage{
snp_cor(
  Gna,
  ind.row = rows_along(Gna),
  ind.col = cols_along(Gna),
  size = 500,
  alpha = 1,
  thr_r2 = 0,
  fill.diag = TRUE,
  infos.pos = NULL,
  ncores = 1
)

bed_cor(
  obj.bed,
  ind.row = rows_along(obj.bed),
  ind.col = cols_along(obj.bed),
  size = 500,
  alpha = 1,
  thr_r2 = 0,
  fill.diag = TRUE,
  infos.pos = NULL,
  ncores = 1
)

snp_cor_with_cov(
  G,
  covar.row,
  ind.row = rows_along(G),
  ind.col = cols_along(G),
  size = 500,
  alpha = 1,
  thr_r2 = 0,
  fill.diag = TRUE,
  infos.pos = NULL,
  ncores = 1
)
}
\arguments{
\item{Gna}{A \link[=FBM.code256-class]{FBM.code256}
(typically \verb{<bigSNP>$genotypes}).\cr
You can have missing values in these data.}

\item{ind.row}{An optional vector of the row indices (individuals) that
are used. If not specified, all rows are used.\cr
\strong{Don't use negative indices.}}

\item{ind.col}{An optional vector of the column indices (SNPs) that are used.
If not specified, all columns are used.\cr
\strong{Don't use negative indices.}}

\item{size}{For one SNP, window size around this SNP to compute correlations.
Default is \code{500}. If not providing \code{infos.pos} (\code{NULL}, the default), this is
a window in number of SNPs, otherwise it is a window in kb (genetic distance).}

\item{alpha}{Type-I error for testing correlations.
Default is \code{1} (no threshold is applied).}

\item{thr_r2}{Threshold to apply on squared correlations. Default is \code{0}.}

\item{fill.diag}{Whether to fill the diagonal with 1s (the default)
or to keep it as 0s.}

\item{infos.pos}{Vector of integers specifying the physical position
on a chromosome (in base pairs) of each SNP.\cr
Typically \verb{<bigSNP>$map$physical.pos}.}

\item{ncores}{Number of cores used. Default doesn't use parallelism.
You may use \link{nb_cores}.}

\item{obj.bed}{Object of type \link{bed}, which is the mapping of some bed file.
Use \code{obj.bed <- bed(bedfile)} to get this object.}

\item{G}{A \link[=FBM.code256-class]{FBM.code256}
(typically \verb{<bigSNP>$genotypes}).\cr
\strong{You shouldn't have missing values.} Also, remember to do quality control,
e.g. some algorithms in this package won't work if you use SNPs with 0 MAF.}

\item{covar.row}{Matrix of covariables (corresponding to \code{ind.row}) to adjust
for when computing correlations. You can use \code{\link[bigstatsr:covar_from_df]{bigstatsr::covar_from_df()}}
to convert from a data frame.}
}
\value{
The (Pearson) correlation matrix. This is a sparse symmetric matrix.
}
\description{
Get significant (Pearson) correlations between nearby SNPs of the same chromosome
(p-values are computed using a two-sided t-test).
}
\examples{
test <- snp_attachExtdata()
G <- test$genotypes

corr <- snp_cor(G, ind.col = 1:1000)
corr[1:10, 1:10]

# Sparsity
length(corr@x) / length(corr)

}
