% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr-thr.R
\name{snp_cor_extendedThr}
\alias{snp_cor_extendedThr}
\title{Extended thresholded correlation matrix}
\usage{
snp_cor_extendedThr(
  Gna,
  thr_r2,
  infos.pos,
  size,
  ind.row = rows_along(Gna),
  ind.col = cols_along(Gna),
  ncores = 1
)
}
\arguments{
\item{Gna}{A \link[=FBM.code256-class]{FBM.code256}
(typically \verb{<bigSNP>$genotypes}).\cr
You can have missing values in these data.}

\item{thr_r2}{Threshold to apply on squared correlations.}

\item{infos.pos}{Vector of integers specifying the physical position
on a chromosome (in base pairs) of each SNP.\cr
Typically \verb{<bigSNP>$map$physical.pos}.}

\item{size}{For one SNP, window size around this SNP to compute correlations.
This is internally multiplied by 1000 (e.g. \code{500} means 500 kb).}

\item{ind.row}{An optional vector of the row indices (individuals) that
are used. If not specified, all rows are used.\cr
\strong{Don't use negative indices.}}

\item{ind.col}{An optional vector of the column indices (SNPs) that are used.
If not specified, all columns are used.\cr
\strong{Don't use negative indices.}}

\item{ncores}{Number of cores used. Default doesn't use parallelism.
You may use \code{\link[=nb_cores]{nb_cores()}}.}
}
\value{
A sparse symmetric matrix.
}
\description{
Get all pairwise correlations \code{r(i,j)} for which there exists k such that
\code{r2(i,k) >= thr_r2} and \code{r2(j,k) >= thr_r2}. For other indices, 0s are returned.
}
\examples{
test <- snp_attachExtdata()
POS <- test$map$physical.pos
G <- test$genotypes
cor(G[, 1:5])
snp_cor_extendedThr(G, thr_r2 = 0.002, ind.col = 1:5,
                  infos.pos = POS[1:5], size = 500)

}
