% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRS.R
\name{snp_cvPRS}
\alias{snp_cvPRS}
\title{Title}
\usage{
snp_cvPRS(x, ind.train, ind.test, covar.train = NULL, method = "auto",
  thr.list = NULL, ncores = 1, K = 10, clumping = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
