% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRS.R
\name{snp_cvPRS_fast}
\alias{snp_cvPRS_fast}
\title{Title}
\usage{
snp_cvPRS_fast(x, ind.train, ind.test, covar.train = NULL, method = "auto",
  thr.list = 0, ncores = 1, K = 10, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
