% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double.R
\name{snp_double}
\alias{snp_double}
\title{Double a genotype matrix}
\usage{
snp_double(x)
}
\arguments{
\item{x}{A \link[=bigSNP-class]{bigSNP}.}
}
\value{
The new genotype matrix, as a \code{big.matrix}.
}
\description{
Recode each column of a genotype matrix to be two columns, one encoding a
recessive status and the other a dominant status.
}
