% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake.R
\name{snp_fake}
\alias{snp_fake}
\title{Fake a "bigSNP"}
\usage{
snp_fake(n, m)
}
\arguments{
\item{n}{Number of individuals.}

\item{m}{Number of SNPs.}
}
\value{
A new temporary \code{bigSNP} object representing \code{n} individuals and
\code{m} SNPs. The genotype \code{big.matrix} is initialized with missing values.
}
\description{
Fake a "bigSNP"
}
\examples{
N <- 5
M <- 12
test <- snp_fake(N, M)
str(test)

# The genotype `big.matrix` is initialized with missing values
X <- test$genotypes
all(is.na(X[,]))

# Modify the genotype `big.matrix`
X[] <- sample(c(0:2, NA), size = length(X), replace = TRUE)
print(X[,])
}
\keyword{internal}
