% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man-qq-gc.R
\name{snp_gc}
\alias{snp_gc}
\title{Genomic Control}
\usage{
snp_gc(gwas)
}
\arguments{
\item{gwas}{}
}
\description{
Genomic Control
}
\references{
Devlin, B., & Roeder, K. (1999).
Genomic control for association studies.
Biometrics, 55(4), 997-1004.
}
