% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{snp_gene}
\alias{snp_gene}
\title{Get genes}
\usage{
snp_gene(rsid, ncores = 1)
}
\arguments{
\item{rsid}{A character vector of 'rs' ID of SNPs to investigate.}

\item{ncores}{Number of cores used. Default doesn't use parallelism.
You may use \link{nb_cores}.}
}
\value{
A character vector of genes in the form \code{"<name>:<ID>". Note that there can be multiple genes per SNP (separated by a comma), or none (}NA`).
}
\description{
Get genes associated with SNPs.
}
\examples{
rsid <- c("rs3934834", "rs3737728", "rs6687776", "rs9651273", "rs4970405",
          "rs12726255", "rs2298217", "rs4970362", "rs9660710", "rs4970420")
snp_gene(rsid)
}
