% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveModifs.R
\name{snp_getPops}
\alias{snp_getPops}
\title{Get population infos}
\usage{
snp_getPops(x, pop.files, col.sample.ID, col.family.ID, ...)
}
\arguments{
\item{pop.files}{Character vector of file names where to
find the population of the individuals of the study.}

\item{col.sample.ID}{Index of the column containing the
sample IDs to match with those of the study.}

\item{col.family.ID}{Index of the column containt the populations.}

\item{...}{Any additional parameter to pass to \link[data.table:fread]{fread}.}
}
\value{
The modified \code{bigSNP} (modifications are saved in .rds).
}
\description{
Get populations of individuals by matching from an external file and
\strong{overwrite} the slot \strong{\code{family.ID}} of the slot \code{fam}.
}
\examples{
test <- snp_readExample()

# Just after reading
print(rle(test$fam$family.ID))

# Get populations clusters from external files
files <- system.file("extdata", paste0("cluster", 1:3), package = "bigsnpr")
print(data.table::fread(files[1]))
test <- snp_getPops(test, pop.files = files, col.sample.ID = 2,
                    col.family.ID = 3)
print(rle(test$fam$family.ID))
}

