% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{snp_impute}
\alias{snp_impute}
\title{Imputation}
\usage{
snp_impute(x, perc.train = 0.7, nrounds = 20, max_depth = 3,
  breaks = c(0, trunc(n/c(1000, 200, 100))), sizes = c(10, 20, 30, 50),
  ncores = 1, verbose = FALSE)
}
\arguments{
\item{x}{A \link[=bigSNP-class]{bigSNP}.}

\item{perc.train}{Percentage of individuals used for the training
(the rest is used to assess the error of imputation by SNP).}

\item{nrounds}{The number of iterations (trees) in
\link[xgboost:xgboost]{xgboost}. Default is \code{20}.}

\item{max_depth}{maximum depth of a tree. Default is \code{3}.}

\item{breaks}{Break points of the number of missing values by SNPs
to which we change the number of predictors to be used. The more missing
values there are, the more complex is the model used.}

\item{sizes}{Radius of how many predictors enter the model.}

\item{ncores}{Number of cores used. Default doesn't use parallelism.}

\item{verbose}{Print progress? Default is \code{FALSE}.}
}
\value{
The new \link{bigSNP} object with a slot \code{imputation} which is
a \code{data.frame} with 2 columns:
\itemize{
\item the number of missing values by SNP,
\item the estimated error of imputation by SNP.
}
}
\description{
Imputation via local XGBoost models.
}
\details{
This implementation aims at imputing with good accuracy while being fast.
For instance, \code{nrounds} and \code{max_depth} are voluntarily chosen low.
One could use for example \code{nrounds = 50} and \code{max_depth = 5} to improve
accuracy a bit, at the expense of computation time.
\cr
Moreover, \code{breaks} and \code{sizes} are used to dynamically choose the number
of predictors to use more complex models when there are more missing values.
To use only complex models, one could use \code{breaks = 0} and \code{sizes = 100},
increasing computation time.
}
