% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{snp_imputeCV}
\alias{snp_imputeCV}
\title{Title}
\usage{
snp_imputeCV(x, nrounds = 20, max_depth = 3, breaks = c(0, trunc(n/c(1000,
  200, 100))), sizes = c(10, 20, 30, 50), Kfolds = c(2, 3, 5, 8),
  ncores = 1, verbose = FALSE)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
