% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsc_cross_trait.R
\name{snp_ldsc_rg}
\alias{snp_ldsc_rg}
\title{Cross-Trait LD score regression}
\usage{
snp_ldsc_rg(
  ld_score,
  ld_size,
  z1,
  z2,
  sample_size_1,
  sample_size_2,
  blocks = 200,
  intercept = NULL,
  step1_chisq_max = 30,
  chi2_thr2 = Inf,
  ncores = 1
)
}
\arguments{
\item{ld_score}{Vector of LD scores.}

\item{ld_size}{Number of variants used to compute \code{ld_score}.}

\item{z1}{Vector of z-scores for trait 1.}

\item{z2}{Vector of z-scores for trait 2.}

\item{sample_size_1}{Sample size of GWAS for trait 1.
Possibly a vector, or just a single value.}

\item{sample_size_2}{Sample size of GWAS for trait 2.
Possibly a vector, or just a single value.}

\item{blocks}{Either a single number specifying the number of blocks,
or a vector of integers specifying the block number of each \code{chi2} value.
Default is \code{200} for \code{snp_ldsc()}, dividing into 200 blocks of approximately
equal size. \code{NULL} can also be used to skip estimating standard errors,
which is the default for \code{snp_ldsc2()}.}

\item{intercept}{You can constrain the intercept to some value (e.g. 0).
Default is \code{NULL} (the intercept is estimated).
Use a value of 0 if you are sure there is no overlap between GWAS samples.}

\item{step1_chisq_max}{Threshold on \code{chi2} in step 1. Default is \code{30}.}

\item{chi2_thr2}{Threshold on \code{chi2} in step 2. Default is \code{Inf} (none).}

\item{ncores}{Number of cores used. Default doesn't use parallelism.
You may use \link{nb_cores}.}
}
\value{
Vector of 4 values (or only the first 2 if \code{blocks = NULL}):
\itemize{
\item \verb{[["int"]]}: LDSC regression intercept,
\item \verb{[["int_se"]]}: SE of this intercept,
\item \verb{[["h2"]]}: LDSC regression estimate of (SNP) heritability (also see
\link{coef_to_liab}),
\item \verb{[["h2_se"]]}: SE of this heritability estimate.
}
}
\description{
Cross-Trait LD score regression
}
