% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS.R
\name{snp_logitGWAS}
\alias{snp_logitGWAS}
\title{Title}
\usage{
snp_logitGWAS(x, ind.train = seq(nrow(X)), covar = NULL, ncores = 1,
  tol = 1e-08, maxiter = 100)
}
\arguments{
\item{maxiter}{}
}
\description{
Title
}
