% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqman.R
\name{snp_manhattan}
\alias{snp_manhattan}
\title{Manhattan plot}
\usage{
snp_manhattan(gwas, map, ...)
}
\arguments{
\item{gwas}{A \code{mhtest} object with the p-values associated with each SNP.
Typically, the output of \link{big_univLinReg} or \link{big_univLogReg}.}

\item{map}{The slot \code{map} of a \link[=bigSNP-class]{bigSNP}.}

\item{...}{Arguments passed on to \code{qqman::manhattan}
\describe{
  \item{col}{A character vector indicating which colors to alternate.}
  \item{chrlabs}{A character vector equal to the number of chromosomes
specifying the chromosome labels (e.g., \code{c(1:22, "X", "Y", "MT")}).}
  \item{suggestiveline}{Where to draw a "suggestive" line. Default
-log10(1e-5). Set to FALSE to disable.}
  \item{genomewideline}{Where to draw a "genome-wide sigificant" line. Default
-log10(5e-8). Set to FALSE to disable.}
  \item{highlight}{A character vector of SNPs in your dataset to highlight.
These SNPs should all be in your dataset.}
  \item{logp}{If TRUE, the -log10 of the p-value is plotted. It isn't very
useful to plot raw p-values, but plotting the raw value could be useful for
other genome-wide plots, for example, peak heights, bayes factors, test
statistics, other "scores," etc.}
}}
}
\value{
\code{NULL}. Creates a Manhattan plot.
}
\description{
Creates a manhattan plot based on the package \strong{qqman}.
}
\references{
Devlin, B., & Roeder, K. (1999).
Genomic control for association studies.
Biometrics, 55(4), 997-1004.
}
