% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAX3.R
\name{snp_MAX3}
\alias{snp_MAX3}
\title{MAX3 statistic}
\usage{
snp_MAX3(x, ind.train = seq(nrow(X)))
}
\arguments{
\item{x}{A \link[=bigSNP-class]{bigSNP}.}

\item{ind.train}{An optional vector of the row indices that are used,
for the training part. If not specified, all data are used.}
}
\value{
A named list of \strong{\code{S}} and \strong{\code{pS}} for every column,
which are MAX3 statistics and associated p-values. \strong{P-values are in
fact the minimum of the 3 p-values of each test separately.} One can use
genomic control to rescale these p-values.
}
\description{
Counts the number of 0, 1, 2 by SNP and phenotype (cases/controls)
and then compute the MAX3 statistic.
}
\examples{
set.seed(1)

# constructing a fake genotype big.matrix
a <- big.matrix(10, 15, type = "char")
a[] <- sample(c(0, 1, 2), 150, TRUE)
print(a[,])

# constructing a fake incomplete bigSNP with 10 individuals and 15 SNPs
# where the 5 first individuals are cases and the 5 last are controls.
fake <- list()
class(fake) <- "bigSNP"
fake$genotypes <- a
fake$fam$affection <- c(rep(1, 5), rep(-1, 5))

# Get MAX3 statistics
print(snp_MAX3(fake))

}

