% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-software.R
\name{snp_plinkIBDQC}
\alias{snp_plinkIBDQC}
\title{Identity-by-descent}
\usage{
snp_plinkIBDQC(plink.path, bedfile.in, bedfile.out = NULL, pi.hat = 0.08,
  ncores = 1, pruning.args = c(100, 0.2), extra.options = "")
}
\arguments{
\item{plink.path}{Path to the executable of PLINK 1.9.}

\item{bedfile.in}{Path to the input bedfile.}

\item{bedfile.out}{Path to the output bedfile. Default is created by
appending \code{"_norel"} to \code{prefix.in} (\code{bedfile.in} without extension).}

\item{pi.hat}{PI_HAT value threshold for individuals (the second)
to be excluded. Default is \code{0.08}.}

\item{ncores}{Number of cores to be used. Default is \code{1}. An usually good
value for this parameter is \code{ncores = parallel::detectCores() - 1}.}

\item{pruning.args}{A vector of 2 pruning parameters, respectively
the window size (in variant count) and the pairwise $r^2$ threshold
(the step size is fixed to 1). Default is \code{c(100, 0.2)}.}

\item{extra.options}{Other options to be passed to PLINK as a string. More
options can be found at \url{https://www.cog-genomics.org/plink/1.9/ibd}.}
}
\value{
The path of the new bedfile.
If no sample is filter, no new bed/bim/fam files are created and
then the path of the input bedfile is returned.
}
\description{
Quality Control based on Identity-by-descent (IBD) computed by
\href{https://www.cog-genomics.org/plink2}{PLINK 1.9}
using its method-of-moments.
}
\examples{
\dontrun{
# On MY computer, I can do:
plink <- "../plink_linux_x86_64/plink"
bedfile <- "../POPRES_data/POPRES_allchr_QC.bed"
test <- snp_plinkIBDQC(plink, bedfile, ncores = 4)}
}
\references{
Purcell, Shaun, Benjamin Neale, Kathe Todd-Brown, Lori Thomas,
Manuel A R Ferreira, David Bender, Julian Maller, et al. 2007.
\emph{PLINK: a tool set for whole-genome association and population-based linkage
analyses.} American Journal of Human Genetics 81 (3). Elsevier: 559–75.
\url{http://dx.doi.org/10.1086/519795}.

Chang, Christopher C, Carson C Chow, Laurent CAM Tellier,
Shashaank Vattikuti, Shaun M Purcell, and James J Lee. 2015.
\emph{Second-generation PLINK: rising to the challenge of larger and richer
datasets.} GigaScience 4 (1): 7.
\url{http://dx.doi.org/10.1186/s13742-015-0047-8}.
}
\seealso{
\link{snp_plinkQC}
}
