% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRS.R
\name{snp_PRS}
\alias{snp_PRS}
\title{PRS}
\usage{
snp_PRS(x, betas, ind.test, lpS = NULL, thr.list = 0, ...)
}
\arguments{
\item{x}{A \link[=bigSNP-class]{bigSNP}.}

\item{betas}{Numeric vector of weights associated with each SNP. You may
want to see \link[bigstatsr:big_univRegLin]{big_univRegLin} or
\link[bigstatsr:big_univRegLog]{big_univRegLog}.}

\item{ind.test}{The individuals on whom to project the scores.}

\item{lpS}{Numeric vector of \code{-log10(p.value)} associated with \code{betas}.
Default doesn't use thresholding.}

\item{thr.list}{Threshold vector on \code{lpS} at which SNPs are excluded if
they are not significant enough. Default doesn't use thresholding.}

\item{...}{Arguments passed on to \code{snp_clumping}
\describe{
  \item{S}{A vector of column statistics which express the importance
of each SNP (the more important is the SNP, the greater should be
the corresponding statistic).
For example, if \code{S} follows the standard normal distribution,
you should probably use \code{abs(S)} instead.}
  \item{size}{This parameter should be adjusted with respect to the number of SNPs.
\itemize{
\item for clumping: \strong{Radius} of the window's size for the LD evaluations.
Default is \code{1000} (I use this for a chip of 500K SNPs).
\item for pruning: \strong{Diameter} of the window's size for the LD evaluations.
Default is \code{50} (as in PLINK 1.07).
}}
  \item{thr.r2}{Threshold over the squared correlation between two SNPs.
Default is \code{0.5}.}
  \item{ind.train}{An optional vector of the row indices that are used,
for the training part. If not specified, all rows are used.}
  \item{ncores}{Number of cores used. Default doesn't use parallelism.}
}}
}
\value{
A matrix of scores, where rows correspond to \code{ind.test} and
columns correspond to \code{thr.list}.
}
\description{
Polygenic Risk Scores with possible clumping and thresholding.
}
