% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning.R
\name{snp_prune}
\alias{snp_prune}
\title{LD pruning}
\usage{
snp_prune(x, ind.train = seq(nrow(X)), size = 50, thr.corr = 0.2,
  exclude = NULL, ncores = 1)
}
\arguments{
\item{x}{A \link[=bigSNP-class]{bigSNP}.}

\item{ind.train}{An optional vector of the row indices that are used,
for the training part. If not specified, all data are used.}

\item{size}{Radius of the window's size for the LD evaluations.
Default is \code{500} for clumping and \code{50} for pruning. This should be
adjusted for different number of SNPs (this corresponds to the defaults
I use for a chip of 500K SNPs).}

\item{thr.corr}{Threshold on the correlation between two SNPs.
Default is \code{0.2}.}

\item{exclude}{Vector of indices of SNPs to exclude anyway. For example,
can be used to exclude long-range LD regions (see Price2008).}

\item{ncores}{Number or cores used. Default doesn't use parallelism.}
}
\description{
LD pruning (and thresholding) for a \code{bigSNP}.
Similar to --indep-pairwise size 1 thr.corr
}
\references{
Price AL, Weale ME, Patterson N, et al.
Long-Range LD Can Confound Genome Scans in Admixed Populations.
Am J Hum Genet. 2008;83(1):132-135.
\url{http://dx.doi.org/10.1016/j.ajhg.2008.06.005}.
}

