% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning.R
\name{snp_pruning}
\alias{snp_pruning}
\title{LD pruning}
\usage{
snp_pruning(x, ind.train = seq(nrow(X)), size = 50, thr.corr = 0.2,
  exclude = NULL, ncores = 1)
}
\arguments{
\item{x}{A \link[=bigSNP-class]{bigSNP}.}

\item{ind.train}{An optional vector of the row indices that are used,
for the training part. If not specified, all data are used.}

\item{size}{\strong{Diameter} of the window's size for the LD evaluations.
Default is \code{50} for pruning (default in PLINK) and should be adjusted
for different number of SNPs.}

\item{thr.corr}{Threshold on the correlation between two SNPs.
Default is \code{0.2}.}

\item{exclude}{Vector of indices of SNPs to exclude anyway. For example,
can be used to exclude long-range LD regions (see Price2008).}

\item{ncores}{Number or cores used. Default doesn't use parallelism.}
}
\description{
LD pruning (and thresholding) for a \code{bigSNP}.
Similar to \code{--indep-pairwise size 1 thr.corr} (\code{step} is fixed to 1).
}
\references{
Purcell, S., Neale, B., Todd-Brown, K., Thomas, L., Ferreira,
M., & Bender, D. et al. (2007). PLINK: A Tool Set for
Whole-Genome Association and Population-Based Linkage Analyses.
The American Journal Of Human Genetics, 81(3), 559-575.
\url{http://dx.doi.org/10.1086/519795}.
}

