% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-sumstats.R
\name{snp_qc_sumstats}
\alias{snp_qc_sumstats}
\title{Quality Control of summary statistics}
\usage{
snp_qc_sumstats(
  z_sumstats,
  corr,
  thr_highld = 0.05,
  num_highld = 20,
  prop_eig = 1,
  max_run = 0.1 * length(z_sumstats),
  p_val_thr = 5e-08,
  print_info = FALSE,
  ncores = 1
)
}
\arguments{
\item{z_sumstats}{z_sumstats from GWAS summary statistics}

\item{corr}{LD matrix}

\item{thr_highld}{Threshold for highly correlated variants, \code{0.05} by default}

\item{num_highld}{Minimum number of highly correlated variants, \code{20} by default}

\item{prop_eig}{Proportion of eigenvalues to attain, \code{1} by default}

\item{max_run}{Maximum round of iterations to run.}

\item{p_val_thr}{P-value Threshold for QC, \code{5e-8} by default}

\item{print_info}{Print_info or not, \code{FALSE} by default}

\item{ncores}{Number of cores to run in parallel}
}
\value{
A dataframe wth three variables of each variants:
\itemize{
\item \verb{$chi2_final}: chi-squared statistics
\item \verb{$p_val}: corresponding p-value
\item \verb{$remove}: corresponding variant being removed or not
}
}
\description{
Quality Control of summary statistics
}
\examples{


obj_bigsnp <- snp_attachExtdata()
G <- obj_bigsnp$genotypes
Z <- big_univLogReg(G, obj_bigsnp$fam$affection - 1, ind.col = 1:500)$score
ind <- which.max(abs(Z))
Z[ind] <- -Z[ind]
ld <- as_SFBM(snp_cor(G, ind.col = 1:500), compact = TRUE)
res <- snp_qc_sumstats(z_sumstats = Z, corr = ld)
cat(res$p_val[ind], res$remove[ind])

}
