% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcimp-sumstats.R
\name{snp_qcimp_sumstats}
\alias{snp_qcimp_sumstats}
\title{Quality control and imputation of summary statistics}
\usage{
snp_qcimp_sumstats(
  corr,
  df_beta,
  sd0,
  thr_highld = 0.2,
  thr_ld = 30,
  removed = rep(FALSE, nrow(df_beta)),
  pval_thr = 5e-08,
  max_run = nrow(df_beta),
  print_iter = FALSE,
  ncores = 1
)
}
\arguments{
\item{corr}{Sparse correlation matrix as an \link[=SFBM-class]{SFBM}.
If \code{corr} is a dsCMatrix or a dgCMatrix, you can use \code{as_SFBM(corr)}.}

\item{df_beta}{A data frame with 3 columns:
\itemize{
\item \verb{$beta}: effect size estimates
\item \verb{$beta_se}: standard errors of effect size estimates
\item \verb{$n_eff}: either GWAS sample size(s) when estimating \code{beta} for a
continuous trait, or in the case of a binary trait, this is
\code{4 / (1 / n_control + 1 / n_case)}; in the case of a meta-analysis, you
should sum the effective sample sizes of each study instead of using the
total numbers of cases and controls, see \doi{10.1016/j.biopsych.2022.05.029};
when using a mixed model, the effective sample size needs to be adjusted
as well, see \doi{10.1016/j.xhgg.2022.100136}.
}}

\item{sd0}{Vector of standard deviations (of genotypes). You can estimate
those using \code{sqrt(2 * af * (1 - af) * INFO)}, where \code{af} are the allele
frequencies and \code{INFO} the imputation info scores.}

\item{thr_highld}{Threshold on squared correlations to decide which variants
to use for imputation. Default is \code{0.2}. This MUST be equal (or larger)
than \code{thr_r2} used in \code{snp_cor_extendedThr()} to compute \code{corr}.}

\item{thr_ld}{Threshold on LD scores for restricting variants in LD to use
for imputation. This is used to avoid redundancy and increase speed.
Default is \code{30}.}

\item{removed}{Logical vector to discard some variants from being used in
the imputation of other variants. However, these variants will be imputed.}

\item{pval_thr}{P-value threshold for QCing. Default is \code{5e-8}.}

\item{max_run}{Maximum number of iterations (and therefore the maximum number
of variants that can be removed). Default is the number of input variants.}

\item{print_iter}{Whether to print the iteration numbers. Default is \code{FALSE}.}

\item{ncores}{Number of cores used. Default doesn't use parallelism.
You may use \code{\link[=nb_cores]{nb_cores()}}.}
}
\value{
A tibble (data frame) with 8 variables for each variants:
\itemize{
\item \verb{$beta_imp}: imputed effect sizes,
\item \verb{$beta_se_imp}: imputed standard errors (corresponding to \verb{$n_eff_imp}),
\item \verb{$n_eff_imp}: imputed sample sizes,
\item \verb{$r2_imp}: multiple $R^2$ from imputation,
\item \verb{$r2_max}: maximum $r^2$ with variants used for imputing,
\item \verb{$chi2_qc}: the chi-squared statistics,
\item \verb{$pval_qc}: the corresponding p-values (df = 1),
\item \verb{$rm_qc}: whether the p-values are significant (or \code{removed}).
}
}
\description{
Quality control of summary statistics, by comparing z-scores with z-scores
imputed using nearby correlated variants. Also performs imputation.
}
