% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqman.R
\name{snp_qq}
\alias{snp_qq}
\title{Q-Q plot}
\usage{
snp_qq(gwas, lambdaGC = TRUE, ...)
}
\arguments{
\item{gwas}{A \code{mhtest} object with the p-values associated with each SNP.
Typically, the output of \link{big_univLinReg} or \link{big_univLogReg}.}

\item{lambdaGC}{Add the Genomic Control cofficient to the plot?}

\item{...}{Other arguments passed to plot().}
}
\value{
\code{NULL}. Creates a Q-Q plot.
}
\description{
Creates a quantile-quantile plot from p-values from a GWAS study based on
the package \strong{qqman}.
}
