% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readplink.R
\name{snp_writeBed}
\alias{snp_writeBed}
\title{Write PLINK files from a "bigSNP"}
\usage{
snp_writeBed(x, bedfile)
}
\arguments{
\item{x}{A \link[=bigSNP-class]{bigSNP}.}

\item{bedfile}{Path to file with extension ".bed" to create.}
}
\value{
The input \code{bedfile} path.
}
\description{
Function to write bed/bim/fam files from a \link[=bigSNP-class]{bigSNP}.
}
