% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-QC.R
\name{subset.bigSNP}
\alias{subset.bigSNP}
\title{Subset}
\usage{
\method{subset}{bigSNP}(x, ind.row = rows_along(X), ind.col = cols_along(X),
  ...)
}
\arguments{
\item{x}{A \link[=bigSNP-class]{bigSNP}.}

\item{ind.row}{Indices of the rows (individuals) to keep.
Negative indices \strong{can} be used to exclude row indices.
Default: keep them all.}

\item{ind.col}{Indices of the columns (SNPs) to keep.
Negative indices \strong{can} be used to exclude column indices.
Default: keep them all.}

\item{...}{Not used.}
}
\value{
The path to the RDS file that stores the \code{bigSNP} object.
Note that this function creates two other files which store the
\code{big.matrix} and its descriptor.
}
\description{
A method to get a subset (copy) of a \code{bigSNP}.
This new \code{bigSNP} will also be backed by files (in the same directory).
}
\seealso{
\link[=bigSNP-class]{bigSNP}
}
