% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC.R
\name{auc}
\alias{AucSample}
\alias{AucSampleConf}
\alias{auc}
\title{AUC.}
\usage{
AucSample(pred, y, nsim = 1e+07, seed = NULL, digits = 3)

AucSampleConf(pred, y, nboot = 1e+05, nsim = 10000, seed = NULL,
  digits = 3)
}
\arguments{
\item{pred}{Vector of predictions.}

\item{y}{Vector of true values (for the same individuals).}

\item{nsim}{Number of pairs (case/control) that are compared
in order to approximate the probability.}

\item{seed}{See \code{\link{set.seed}}. Use it for reproducibility.}

\item{digits}{See \code{\link{round}}.}

\item{nboot}{Number of bootstrap samples to evaluate 95\% CI.}
}
\value{
The AUC, a probability, and possibly its 2.5\% and 97.5\% quantiles
(95\% CI).
}
\description{
Compute the Area Under the ROC Curve (AUC) of a predictor
and possibly its 95\% confidence interval.
}
\details{
Other packages provide ways to compute the AUC.
Yet, I found out that when there are equalities, those can
overestimate the true AUC. So, I choose to compute the AUC
through an (accurate) estimation of its statistical definition as a
probability: \deqn{\P{score(x_{case}) > score(x_{control})}}
}
\references{
Tom Fawcett. 2006. An introduction to ROC analysis.
Pattern Recogn. Lett. 27, 8 (June 2006), 861-874.
DOI=\url{http://dx.doi.org/10.1016/j.patrec.2005.10.010}.
}
\seealso{
pROC::\code{\link[pROC]{auc}} \cr
AUC::\code{\link[AUC]{auc}}
}

