% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigsnpr.R
\docType{package}
\name{bigsnpr-package}
\alias{bigsnpr}
\alias{bigsnpr-package}
\title{bigsnpr: Predictive Analysis of Massive SNP Arrays}
\arguments{
\item{X}{A \link[bigmemory:big.matrix-class]{big.matrix}.
You shouldn't have missing values in your data.}

\item{x}{A \link[=bigSNP-class]{bigSNP}.}

\item{has.pheno}{Does \strong{\code{x}} has well-specified phenotypes?
i.e. \code{x$fam$affection} is a vector with exactly two values
(and the value coding for cases is larger than the value coding
for controls). So, in particular, no missing value is allowed.
Default is \code{TRUE}.}

\item{ind.train}{An optional vector of the row indices that are used,
for the training part. If not specified, all data are used.}

\item{block.size}{Maximum number of loci read at once (for all individuals).}

\item{ncores}{Number or cores used. Default doesn't use parallelism.}
}
\description{
Easy-to-use, efficient, flexible and scalable tools
for predictive analysis based on SNP arrays (case-control studies).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/privefl/bigsnpr}
  \item Report bugs at \url{https://github.com/privefl/bigsnpr/issues}
}

}
\author{
\strong{Maintainer}: Florian Privé \email{florian.prive.21@gmail.com}

}

