% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigsnpr.R
\docType{package}
\name{bigsnpr-package}
\alias{bigsnpr}
\alias{bigsnpr-package}
\title{Predictive Analysis of Massive SNP Arrays}
\arguments{
\item{x}{A \link[=bigSNP-class]{bigSNP}.}

\item{ind.train}{An optional vector of the row indices that are used,
for the training part. If not specified, all data are used.}

\item{block.size}{Maximum number of loci read at once (for all individuals).}

\item{ncores}{Number or cores used. Default doesn't use parallelism.}
}
\description{
Easy-to-use, efficient, flexible and scalable tools
for predictive analysis based on SNP arrays (case-control studies).
}

