% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigsnpr.R
\name{bigsnpr-package}
\alias{bigsnpr}
\alias{bigsnpr-package}
\title{R Package for analysis of massive SNP arrays.}
\arguments{
\item{x}{A \code{bigSNP}.}

\item{ind.train}{An optional vector of the row indices that are used,
for the training part.
If not specified, all data are used.}

\item{ncores}{Number or cores used.
Default doesn't use parallelism.}
}
\description{
Easy-to-use, efficient, flexible and scalable
tools for predictive analysis of SNP arrays.
}
\keyword{package}

