% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed-projectPCA.R
\name{download_1000G}
\alias{download_1000G}
\title{Download 1000G}
\usage{
download_1000G(dir, overwrite = FALSE, delete_zip = TRUE)
}
\arguments{
\item{dir}{The directory where to put the downloaded files.}

\item{overwrite}{Whether to overwrite files when downloading and unzipping?
Default is \code{FALSE}.}

\item{delete_zip}{Whether to delete zip after decompressing the file in it?
Default is \code{TRUE}.}
}
\value{
The path of the downloaded bed file.
}
\description{
Download 1000 genomes project (phase 3) data in PLINK bed/bim/fam format,
including 2490 (mostly unrelated) individuals
and ~1.7M SNPs in common with either HapMap3 or the UK Biobank.
}
