% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-software.R
\name{download_plink}
\alias{download_plink}
\alias{download_plink2}
\title{Download PLINK}
\usage{
download_plink(dir = tempdir())

download_plink2(dir = tempdir(), AVX2 = TRUE)
}
\arguments{
\item{dir}{The directory where to put the PLINK executable.
Default is a temporary directory.}

\item{AVX2}{Whether to download the AVX2 version?
Default is \code{TRUE} and possible only for 64 bits architectures.}
}
\value{
The path of the downloaded PLINK executable.
}
\description{
Download PLINK 1.9 from \url{http://www.cog-genomics.org/plink2}.

Download PLINK 2.0 from \url{http://www.cog-genomics.org/plink/2.0/}.
}
