% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-software.R
\name{download_plink}
\alias{download_plink}
\title{Download PLINK 1.9}
\usage{
download_plink(dir = tempdir())
}
\arguments{
\item{dir}{The directory where to put the PLINK executable.
Default is a temporary directory.}
}
\value{
The path of the downloaded PLINK executable.
}
\description{
Download PLINK 1.9 from \url{https://www.cog-genomics.org/plink2}.
}
