% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-sumstats.R
\name{eigen_halfinv}
\alias{eigen_halfinv}
\title{Eigen value decomposition to get sqrt root of inverse of x}
\usage{
eigen_halfinv(x, prop_eig, eig_tol = 1e-04)
}
\arguments{
\item{x}{A symmetric matrix}

\item{prop_eig}{Proportion of eigen values to be attained}

\item{eig_tol}{Min eigenvalue attained}
}
\value{
A pseudo sqrt root of inverse of x
}
\description{
Eigen value decomposition to get sqrt root of inverse of x
}
