% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.bigSNP.R
\name{GBLUP}
\alias{GBLUP}
\title{gBLUP of a "bigSNP".}
\usage{
GBLUP(x, block.size, ind.train, thr.eigval = 0.001, use.Eigen = TRUE)
}
\arguments{
\item{x}{A \code{bigSNP}.}

\item{block.size}{Maximum number of loci read at once (for all individuals).}

\item{ind.train}{An optional vector of the row indices that are used,
for the training part.
If not specified, all data are used.}

\item{thr.eigval}{Threshold to remove eigenvalues to get
stability in the inversion of the matrix.
Default is \code{1e-3}.}

\item{use.Eigen}{Use the \code{Eigen} library to compute \eqn{X X^T}, the default.
If \code{FALSE}, use \code{R}'s \code{tcrossprod}. See details.}
}
\value{
A \code{vector} of predictions.
}
\description{
Genetic Best Linear Unbiased Predictor (gBLUP)
 of a \code{bigSNP}
}
\details{
See \code{\link{BigXYt}}.

Note that for the Eigen decomposition, only \code{R} is
used because is faster (see \href{http://goo.gl/UYJcCw}{stackoverflow}).
If you want a large number of eigenvectors/values, you should
really considerer using Microsoft R Open for speed.
}
\examples{
\dontrun{

  bedfile <- system.file("extdata", "example.bed", package = "bigsnpr")

  # Creating directory for backing files
  if (!dir.exists("backingfiles")) dir.create("backingfiles")

  if (!file.exists("backingfiles/test_doc.bk")) {
    # Reading the bedfile and storing the data
    test <- BedToBig(bedfile, 50, "test_doc")
  } else {
    # Loading it from backing files
    test <- AttachBigSNP("test_doc")
  }

  # create some ramdon class
  test$fam$affection <- sample(c(-1, 1),
                               length(test$fam$affection),
                               replace = TRUE)

  # Get predictions from half the data
  n <- nrow(test$genotypes)
  ind.train <- sort(sample(n, n/2))
  test2 <- GBLUP(test, block.size = 1000, ind.train = ind.train)
  print(AucSampleConf(test2, test$fam$affection[-ind.train], nboot = 1e4))
}
}
\seealso{
\code{\link{bigSNP}}
}

