% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-sumstats.R
\name{impute_z}
\alias{impute_z}
\title{Impute z-score using highly correlated variants}
\usage{
impute_z(id_current, id_high, ld_current_high, z_high, prop_eig)
}
\arguments{
\item{id_current}{current id of variant to be impute}

\item{id_high}{id of highly correlated variants}

\item{ld_current_high}{LD matrix for id_current and id_high}

\item{z_high}{z-scores for highly correlated variants}

\item{prop_eig}{proportion of eigenvalues to be used in eigenvalue decomp}
}
\value{
Imputed z-score, and denominator of chisq statistics
}
\description{
Impute z-score using highly correlated variants
}
