% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-sumstats.R
\name{impute_z}
\alias{impute_z}
\title{Impute z-score using nearby variants}
\usage{
impute_z(ld_current_high, ld_high_high, z_high, prop_eigs)
}
\arguments{
\item{ld_current_high}{LD between variant to impute and its high LD variants.}

\item{ld_high_high}{LD matrix between the high LD variants.}

\item{z_high}{z-scores of \code{id_high} variants.}

\item{prop_eigs}{The proportion of eigen components to keep.}
}
\value{
The imputed z-score, as well as the denominator of the chi-squared
statistic (z - z_imp)^2 / deno.
}
\description{
Impute z-score using nearby variants
}
\examples{
bigsnp <- snp_attachExtdata()
G <- bigsnp$genotypes
Z <- big_univLogReg(G, bigsnp$fam$affection - 1, ind.col = 1:500)$score
ld <- snp_cor(G, ind.col = 1:500)
id_current <- which.max(Matrix::colSums(ld^2))
ld_current <- ld[, id_current]
id_high <- setdiff(which(ld_current^2 > 0.02), id_current)
z_imp <- bigsnpr:::impute_z(ld_current_high = ld_current[id_high],
                            ld_high_high = ld[id_high, id_high],
                            z_high = Z[id_high],
                            prop_eigs = 0.4)
c(Z[id_current], z_imp[1])
chi2 <- (Z[id_current] - z_imp[1])^2 / z_imp[2]
pchisq(chi2, df = 1, lower.tail = FALSE)

}
\references{
Chen, Wenhan, et al. "Improved analyses of GWAS summary statistics
by reducing data heterogeneity and errors." Nature Communications 12.1
(2021): 7117. \doi{10.1038/s41467-021-27438-7}.
}
