% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDpred2.R
\name{snp_ldpred2_inf}
\alias{snp_ldpred2_inf}
\alias{snp_ldpred2_grid}
\alias{snp_ldpred2_auto}
\title{LDpred2}
\usage{
snp_ldpred2_inf(corr, df_beta, h2 = NULL)

snp_ldpred2_grid(
  corr,
  df_beta,
  grid_param,
  burn_in = 100,
  num_iter = 200,
  ncores = 1
)

snp_ldpred2_auto(
  corr,
  df_beta,
  vec_p_init = 0.1,
  h2_init = NULL,
  burn_in = 2000,
  num_iter = 500,
  verbose = FALSE,
  ncores = 1
)
}
\arguments{
\item{corr}{Sparse correlation matrix.}

\item{df_beta}{A data frame with 3 columns:
\itemize{
\item \verb{$beta}: effect size estimates
\item \verb{$beta_se}: standard errors of effect size estimates
\item \verb{$n_eff}: sample size when estimating \code{beta}
(in the case of binary traits, this is \code{4 / (1 / n_control + 1 / n_case)})
}}

\item{h2}{Heritability estimate.
Default is estimated using constrained LD score regression.}

\item{grid_param}{A data frame with 3 columns as a grid of hyper-parameters:
\itemize{
\item \verb{$p}: proportion of causal variants
\item \verb{$h2}: heritability (captured by the variants used)
\item \verb{$sparse}: boolean, whether a sparse model is sought
They can be run in parallel by changing \code{ncores}.
}}

\item{burn_in}{Number of burn-in iterations.}

\item{num_iter}{Number of iterations after burn-in.}

\item{ncores}{Number of cores used. Default doesn't use parallelism.
You may use \link{nb_cores}.}

\item{vec_p_init}{Vector of initial values for p. Default is \code{0.1}.
They can be run in parallel by changing \code{ncores}.}

\item{h2_init}{Heritability estimate for initialization.
Default is estimated using constrained LD score regression.}

\item{verbose}{Whether to print "p // h2" estimates at each iteration.}
}
\value{
\code{snp_ldpred2_inf}: A vector of effects, assuming an infinitesimal model.

\code{snp_ldpred2_grid}: A matrix of effect sizes, one vector (column)
for each row of \code{grid_param}.

\code{snp_ldpred2_auto}: A list of \code{length(vec_p_init)} list(s) with
\itemize{
\item \verb{$beta_est}: vector of effect sizes
\item \verb{$p_est}: estimate of p, the proportion of causal variants
\item \verb{$h2_est}: estimate of the (SNP) heritability
\item \verb{$path_p_est}: full path of p estimates (including burn-in);
useful to check convergence of the iterative algorithm
\item \verb{$path_h2_est}: full path of h2 estimates (including burn-in);
useful to check convergence of the iterative algorithm
}
}
\description{
LDpred2
}
