% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{modif-save}
\alias{modif-save}
\alias{SaveModifs}
\alias{GetPops}
\alias{GetPhenos}
\title{Modify a "bigSNP" and save these modifications.}
\usage{
SaveModifs(x)

GetPops(x, pop.files = NULL, col.sample.ID, col.family.ID, ...)

GetPhenos(x, coded01 = FALSE)
}
\arguments{
\item{x}{A \code{bigSNP}.}

\item{pop.files}{Character vector of file names where to
find the population of the individuals of the study.
Default just copy the slot "family.ID" to the slot "pop".}

\item{col.sample.ID}{Index of the column containing the
sample IDs to match with those of the study.}

\item{col.family.ID}{Index of the column containt the populations.}

\item{...}{Any additional parameter to pass to \code{fread}.}

\item{coded01}{Are the phenotypes coded 0 (unaffected) / 1 (affected)
rather than respectively 1 and 2? Default is \code{FALSE}.}
}
\value{
The (modified) \code{bigSNP}.
}
\description{
Modify a \code{bigSNP} and save these modifications
in the corresponding .rds file. As \code{bigSNP} is an S3 class, you
can add any slot you want to this class, then use \code{SaveModifs} to
save modifications in the corresponding .rds backing file.

\code{GetPops}: Get populations of individuals by matching
from an external file or get them directly from the slot "family.ID" of
the slot "fam".
It will add an slot called "pop" to the slot "fam".

\code{GetPhenos}: Modify phenotypes to be -1 (unaffected),
1 (affected) or NA (missing).
It will add an slot called "pheno" to the slot "fam".
}
\examples{
# Creating directory for backing files
if (!dir.exists("backingfiles")) dir.create("backingfiles")

if (!file.exists("backingfiles/test_doc.bk")) {
  # Reading the bedfile and storing the data
  bedfile <- system.file("extdata", "example.bed", package = "bigsnpr")
  test <- BedToBig(bedfile, 50, "test_doc")
} else {
  # Loading it from backing files
  test <- AttachBigSNP("test_doc")
}

# Just after reading
print(rle(test$fam$family.ID))
print(test$fam$pop) # NULL

# Get populations from slot "family.ID"
test <- GetPops(test)
print(rle(test$fam$pop))

# Get populations clusters from external files
files <- system.file("extdata", paste0("cluster", 1:3), package = "bigsnpr")
test <- GetPops(x = test, pop.files = files, col.sample.ID = 2, col.family.ID = 3)
print(rle(test$fam$pop))
}
\seealso{
\code{\link{rle}}

\code{\link[data.table]{fread}}
}

