% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{modif-save}
\alias{GetPops}
\alias{SaveModifs}
\alias{modif-save}
\title{Modify a "bigSNP" and save these modifications.}
\usage{
SaveModifs(x)

GetPops(x, pop.files, col.sample.ID, col.family.ID, to.save = FALSE)
}
\arguments{
\item{x}{A \code{bigSNP}.}

\item{pop.files}{Character vector of file names where to
find the population of the individuals of the study.}

\item{to.save}{Is the modification to be saved ?}
}
\value{
Nothing.
}
\description{
Modify a \code{bigSNP} and save these modifications
in the corresponding .rds file.
}

