% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readplink.R
\name{readplink}
\alias{readplink}
\alias{BedToBig}
\alias{AttachBigSNP}
\alias{snp_readExample}
\title{Read PLINK files into a "bigSNP".}
\usage{
BedToBig(bedfile, block.size = 1000, backingfile,
  backingpath = "backingfiles")

AttachBigSNP(backingfile, backingpath = "backingfiles", readonly = TRUE)

snp_readExample(backingfile = "test_doc", backingpath = "backingfiles")
}
\arguments{
\item{bedfile}{Path to file with extension .bed. You need the corresponding
.bim and .fam in the same directory.}

\item{backingfile}{The root name for the backing file(s) for the cache of
the resulting object.}

\item{backingpath}{The path to the directory containing the file backing cache.
Default is "backingfiles". It needs to exist (use \link{dir.create}).}

\item{readonly}{Is the \code{big.matrix} read only? Default is \code{TRUE}.}
}
\value{
A \code{bigSNP}.\cr
Reading PLINK files creates
\code{backingfile}.bk, \code{backingfile}.desc and \code{backingfile}.rds
in directory \code{backingpath}.\cr
You shouldn't read from PLINK files more than once.
Instead, use \code{AttachBigSNP}
to load this object in another session from backing files.
}
\description{
Functions to read bed/bim/fam files
into a \code{bigSNP}.\cr
For more information on these formats, please visit
\href{http://pngu.mgh.harvard.edu/~purcell/plink/data.shtml#bed}{PLINK webpage}.
For other formats, please use PLINK to convert them in bedfiles,
which require minimal space to store and are faster to read.
}
\note{
The implementation was originally inspired from the code of
\href{https://github.com/andrewparkermorgan/argyle}{package argyle}.
I did many optimizations. Especially, online reading
into a \code{big.matrix} makes it memory-efficient.
}
\examples{
\dontrun{

  bedfile <- system.file("extdata", "example.bed", package = "bigsnpr")

  # Creating directory for backing files
  if (!dir.exists("backingfiles")) dir.create("backingfiles")

  # Reading the bedfile and storing the data in directory "backingfiles"
  if (!file.exists("backingfiles/test_doc.bk"))
    test <- BedToBig(bedfile, 50, "test_doc")

  # Removing the R object
  rm(test)

  # Loading it from backing files
  test <- AttachBigSNP("test_doc")

  str(test)
  print(dim(test$genotypes))
  print(test$genotypes[1:8, 1:8])
}
}

