% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readplink.R
\name{readplink}
\alias{readplink}
\alias{snp_readBed}
\alias{snp_attach}
\alias{snp_writeBed}
\alias{snp_readExample}
\title{Read|write PLINK files into|from a "bigSNP".}
\usage{
snp_readBed(bedfile, block.size = 1000, backingfile,
  backingpath = "backingfiles")

snp_attach(backingfile, readonly = TRUE)

snp_writeBed(x, bedfile)

snp_readExample(backingfile = "test_doc", backingpath = "backingfiles")
}
\arguments{
\item{bedfile}{Path to file with extension ".bed" to read|create.
When reading, you need the corresponding ".bim" and ".fam"
in the same directory.}

\item{backingfile}{The root name for the backing file(s) for the cache of
the \link[=bigSNP-class]{bigSNP} object.}

\item{backingpath}{The path to the directory containing the file backing cache.
Default is \code{"backingfiles"}.}

\item{readonly}{Is the \code{big.matrix} read only? Default is \code{TRUE}.}

\item{backingfile}{The path of one of the three (".bk", ".desc" or ".rds")
backing files for the cache of the resulting object.}
}
\value{
A \link[=bigSNP-class]{bigSNP}.\cr Reading PLINK files creates
\code{<backingfile>}.bk, \code{<backingfile>}.desc and
\code{<backingfile>}.rds in directory \code{<backingpath>}.\cr
\strong{You shouldn't read from PLINK files more than once.}
Instead, use \code{AttachBigSNP}
to load this object in another session from backing files.
}
\description{
Functions to read|write bed/bim/fam files into|from a \link[=bigSNP-class]{bigSNP}.
\cr For more information on these formats, please visit
\href{http://pngu.mgh.harvard.edu/~purcell/plink/data.shtml#bed}{PLINK webpage}.
For other formats, please use PLINK to convert them in bedfiles,
which require minimal space to store and are faster to read.
}
\note{
The implementation was originally inspired from the code of
\href{https://github.com/andrewparkermorgan/argyle}{package argyle}.
I did many optimizations. Especially, online reading
into a \code{big.matrix} makes it memory-efficient.
}
\examples{
\dontrun{

  bedfile <- system.file("extdata", "example.bed", package = "bigsnpr")

  # Creating directory for backing files
  if (!dir.exists("backingfiles")) dir.create("backingfiles")

  # Reading the bedfile and storing the data in directory "backingfiles"
  if (!file.exists("backingfiles/test_doc.bk"))
    test <- BedToBig(bedfile, 50, "test_doc")

  # Removing the R object
  rm(test)

  # Loading it from backing files
  test <- AttachBigSNP("test_doc")

  str(test)
  print(dim(test$genotypes))
  print(test$genotypes[1:8, 1:8])
}
}
