% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readplink.R
\name{readplink}
\alias{AttachBigSNP}
\alias{BedToBig}
\alias{PedToBig}
\alias{readplink}
\title{Read PLINK files}
\usage{
BedToBig(bedfile, block.size, backingfile, backingpath = "backingfiles")

PedToBig(pedfile, block.size, backingfile, backingpath = "backingfiles")

AttachBigSNP(backingfile, backingpath = "backingfiles", readonly = TRUE)
}
\arguments{
\item{bedfile}{Path to file with extension .bed. You need the corresponding
.bim and .fam in the same directory.}

\item{block.size}{Maximum number of loci read at once (for all individuals).}

\item{backingfile}{The root name for the file(s) for the cache of x.}

\item{backingpath}{The path to the directory containing the file backing cache.
Default is "backingfiles".}

\item{pedfile}{Path to file with extension .ped. You need the corresponding
.map in the same directory.}

\item{readonly}{Is the \code{big.matrix} readonly ? Default is \code{TRUE}.}
}
\value{
A named list of 3 elements:\itemize{
\item genotypes: a filebacked \code{big.matrix} representing genotypes.\cr
Each element is either 0, 1, 2 or NA.
\item fam: a \code{data.table} giving some information on the SNPs.
\item map: a \code{data.table} giving some information on the individuals.
}
}
\description{
Functions to read ped/map or bed/bim/fam files
into a \code{\link[bigmemory]{big.matrix}} (genotypes)
and two \code{\link[data.table]{data.table}} objects
(informations on SNPs and individuals).
For more information on these formats, please visit
\href{http://pngu.mgh.harvard.edu/~purcell/plink/data.shtml#ped}{PLINK webpage}.
}
\note{
The implementation is inspired from the code of
\href{https://github.com/andrewparkermorgan/argyle}{package argyle},
with some optimizations.\cr
Especially, online reading into a \code{big.matrix} makes it memory-efficient.
}
\section{Warnings}{

\itemize{
\item Never \code{save} the result as an \code{R} object,
your session will crash when loading it.\cr
To load the object in another session, use \code{AttachBigSNP}.
\item Prefer using bedfiles than pedfiles because
they require minimal space to store and are faster to read.
}
}
\examples{
\dontrun{

bedfile <- system.file("extdata", "example.bed", package = "mypack")

if (!file.exists("backingfiles"))
 dir.create("backingfiles")
if (file.exists("backingfiles/test_doc"))
 file.remove("backingfiles/test_doc")
if (file.exists("backingfiles/test_doc.desc"))
 file.remove("backingfiles/test_doc.desc")
if (file.exists("backingfiles/test_doc.rds"))
 file.remove("backingfiles/test_doc.rds")

# Reading the bedfile and storing the data in directory "backingfiles"
test <- BedToBig(bedfile, 50, "test_doc")

# Removing the R object
rm(test)

# Loading it from backing files
test <- AttachBigSNP("test_doc")

str(test)
print(dim(test$genotypes))
print(test$genotypes[1:8, 1:8])
}
}

