% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readplink.R
\name{readplink}
\alias{AttachBigSNP}
\alias{BedToBig}
\alias{PedToBig}
\alias{readplink}
\title{Read PLINK files into a "bigSNP".}
\usage{
BedToBig(bedfile, block.size, backingfile, backingpath = "backingfiles")

PedToBig(pedfile, block.size, backingfile, backingpath = "backingfiles")

AttachBigSNP(backingfile, backingpath = "backingfiles", readonly = TRUE)
}
\arguments{
\item{bedfile}{Path to file with extension .bed. You need the corresponding
.bim and .fam in the same directory.}

\item{block.size}{Maximum number of loci read at once (for all individuals).}

\item{backingfile}{The root name for the backing file(s) for the cache of
the resulting object.}

\item{backingpath}{The path to the directory containing the file backing cache.
Default is "backingfiles". It needs to exist.}

\item{pedfile}{Path to file with extension .ped. You need the corresponding
.map in the same directory.}

\item{readonly}{Is the \code{big.matrix} read only ? Default is \code{TRUE}.}
}
\value{
A \code{bigSNP}.\cr
Reading PLINK files creates
\code{backingfile}.bk, \code{backingfile}.desc and \code{backingfile}.rds
in directory \code{backingpath}.\cr
You shouldn't read from PLINK files more than once.
Instead, use \code{AttachBigSNP}
to load this object in another session from backing files.
}
\description{
Functions to read ped/map or bed/bim/fam files
into a \code{\link{bigSNP}}.\cr
For more information on these formats, please visit
\href{http://pngu.mgh.harvard.edu/~purcell/plink/data.shtml#ped}{PLINK webpage}.
Prefer using bedfiles than pedfiles because
they require minimal space to store and are faster to read.
}
\note{
The implementation is inspired from the code of
\href{https://github.com/andrewparkermorgan/argyle}{package argyle},
with some optimizations.\cr
Especially, online reading into a \code{big.matrix} makes it memory-efficient.
}
\examples{
\dontrun{

bedfile <- system.file("extdata", "example.bed", package = "bigsnpr")

if (!dir.exists("backingfiles"))
 dir.create("backingfiles")
if (file.exists("backingfiles/test_doc"))
 file.remove("backingfiles/test_doc")
if (file.exists("backingfiles/test_doc.desc"))
 file.remove("backingfiles/test_doc.desc")
if (file.exists("backingfiles/test_doc.rds"))
 file.remove("backingfiles/test_doc.rds")

# Reading the bedfile and storing the data in directory "backingfiles"
test <- BedToBig(bedfile, 50, "test_doc")

# Removing the R object
rm(test)

# Loading it from backing files
test <- AttachBigSNP("test_doc")

str(test)
print(dim(test$genotypes))
print(test$genotypes[1:8, 1:8])
}
}
\seealso{
\code{\link{dir.create}}
}

