% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDpred2.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_SFBM}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bigsparser}{\code{\link[bigsparser:SFBM-class]{as_SFBM}}}
}}

