% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readplink.R
\name{snp_attach}
\alias{snp_attach}
\title{Attach a "bigSNP" from backing files}
\usage{
snp_attach(rdsfile)
}
\arguments{
\item{rdsfile}{The path of the ".rds" which stores the \code{bigSNP} object.}
}
\value{
The \code{bigSNP} object.
}
\description{
Load a \link[=bigSNP-class]{bigSNP} from backing files into R.
}
\details{
This is often just a call to \link{readRDS}. But it also checks if you have moved
the two (".bk" and ".rds") backing files to another directory.
}
\examples{
bedfile <- system.file("extdata", "example.bed", package = "bigsnpr")
print(bedfile)

# Reading the bedfile and storing the data in directory "backingfiles"
PATH <- "backingfiles/test_doc.bk"
if (!file.exists(PATH))
  PATH <- snp_readBed(bedfile, backingfile = "test_doc")

# Loading the data from backing files
test <- snp_attach(PATH)

str(test)
print(dim(test$genotypes))
print(test$genotypes[1:8, 1:8])
}
