% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readplink.R
\name{snp_attach}
\alias{snp_attach}
\title{Attach a "bigSNP" from backing files}
\usage{
snp_attach(backingfile, readonly = TRUE)
}
\arguments{
\item{backingfile}{The path of one of the three (".bk", ".desc" or ".rds")
backing files for the cache of the "bigSNP" object.}

\item{readonly}{Is the \code{big.matrix} read only? Default is \code{TRUE}.}
}
\value{
The "bigSNP" object.
}
\description{
Load a \link[=bigSNP-class]{bigSNP} from backing files into R.
}
\examples{
bedfile <- system.file("extdata", "example.bed", package = "bigsnpr")
print(bedfile)

# Reading the bedfile and storing the data in directory "backingfiles"
PATH <- "backingfiles/test_doc.bk"
if (!file.exists(PATH))
  PATH <- snp_readBed(bedfile, backingfile = "test_doc")

# Loading the data from backing files
test <- snp_attach(PATH)

str(test)
print(dim(test$genotypes))
print(test$genotypes[1:8, 1:8])
}
