% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readplink.R
\name{snp_attachExtdata}
\alias{snp_attachExtdata}
\title{Attach an "bigSNP" for examples and tests}
\usage{
snp_attachExtdata(backingfile = "test_doc", backingpath = "backingfiles")
}
\arguments{
\item{backingfile}{The root name for the backing file(s) for the cache of
the \link[=bigSNP-class]{bigSNP} object.}

\item{backingpath}{The path to the directory containing the backing files.
Default is \code{"backingfiles"}.}
}
\value{
The example "bigSNP".
}
\description{
Attach an "bigSNP" for examples and tests
}
