% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-plink.R
\name{snp_attachExtdata}
\alias{snp_attachExtdata}
\title{Attach a "bigSNP" for examples and tests}
\usage{
snp_attachExtdata(bedfile = c("example.bed", "example-missing.bed"))
}
\arguments{
\item{bedfile}{Name of one example bed file. Either
\itemize{
\item \code{"example.bed"} (the default),
\item \code{"example-missing.bed"}.
}}
}
\value{
The example "bigSNP", filebacked in the "/tmp/" directory.
}
\description{
Attach a "bigSNP" for examples and tests
}
