% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{snp_cor}
\alias{snp_cor}
\title{Correlation}
\usage{
snp_cor(G, ind.row = rows_along(G), ind.col = cols_along(G), size = 500,
  alpha = 0.05, fill.diag = TRUE)
}
\arguments{
\item{G}{A \link[=BM.code.descriptor-class]{BM.code.descriptor}.
Typically \code{<bigSNP>$genotypes}.\cr
\strong{You shouldn't have missing values in your data or SNPs with 0 MAF.}}

\item{ind.row}{An optional vector of the row indices (individuals) that
are used. If not specified, all rows are used.\cr
\strong{Don't use negative indices.}}

\item{ind.col}{An optional vector of the column indices (SNPs) that are used.
If not specified, all columns are used.\cr
\strong{Don't use negative indices.}}

\item{size}{For one SNP, number of SNPs at its left and its right to
be tested for being correlated with this particular SNP.}

\item{alpha}{Type-I error for testing correlations.}

\item{fill.diag}{Whether to fill the diagonal with 1s (the default)
or to keep it as 0s.}
}
\value{
The correlation matrix. This is a sparse symmetric matrix.
}
\description{
Get significant correlations between nearby SNPs.
}
\examples{
test <- snp_attachExtdata()

corr <- snp_cor(test$genotypes)
corr[1:10, 1:10]

length(corr@x) / length(corr)

}
