% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake.R
\name{snp_fake}
\alias{snp_fake}
\title{Fake a "bigSNP"}
\usage{
snp_fake(n, m)
}
\arguments{
\item{n}{Number of individuals.}

\item{m}{Number of SNPs.}
}
\value{
A new temporary \code{bigSNP} object representing \code{n} individuals
and \code{m} SNPs. The genotype Filebacked Big Matrix is initialized
with missing values.
}
\description{
Fake a "bigSNP"
}
\examples{
N <- 5
M <- 12
(test <- snp_fake(N, M))

# The genotype Filebackeg Big Matrix is initialized with missing values
G <- test$genotypes
X[,]

# Modify the genotype `big.matrix`
X[] <- sample(as.raw(0:3), size = length(X), replace = TRUE)
X[,]

rm(X)

}
\keyword{internal}
