% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake.R
\name{snp_fake}
\alias{snp_fake}
\title{Fake a "bigSNP"}
\usage{
snp_fake(n, m)
}
\arguments{
\item{n}{Number of individuals.}

\item{m}{Number of SNPs.}
}
\value{
A new temporary \code{bigSNP} object representing \code{n} individuals
and \code{m} SNPs. The genotype Filebacked Big Matrix is initialized
with missing values.
}
\description{
Fake a "bigSNP"
}
\examples{
(test <- snp_fake(5, 12))

# The genotype Filebackeg Big Matrix is initialized with missing values
G <- test$genotypes
G[]

# Modify the genotype `big.matrix`
G[] <- sample(as.raw(0:3), size = length(G), replace = TRUE)
G[]

}
\keyword{internal}
