% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{snp_fastImpute}
\alias{snp_fastImpute}
\title{Fast imputation}
\usage{
snp_fastImpute(G, infos.chr, alpha = 0.05, size = 100, seed = NA,
  ncores = 1)
}
\arguments{
\item{G}{A \link[=BM.code.descriptor-class]{BM.code.descriptor}.
Typically \code{<bigSNP>$genotypes}.\cr
\strong{You shouldn't have missing values in your data or SNPs with 0 MAF.}}

\item{infos.chr}{Vector of integers specifying each SNP's chromosome.\cr
Typically \code{<bigSNP>$map$chromosome}.}

\item{alpha}{Type-I error for testing correlations.}

\item{size}{Number of neighbor SNP to be possibly included in the model
imputing this particular SNP.}

\item{seed}{An integer, for reproducibility.}

\item{ncores}{Number of cores used. Default doesn't use parallelism.}
}
\value{
A \code{data.frame} with
\itemize{
\item the proportion of missing values by SNP,
\item the estimated proportion of imputaton errors by SNP.
}
}
\description{
Fast imputation algorithm based on local XGBoost models.
}
