% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{snp_imputeCV2}
\alias{snp_imputeCV2}
\title{Title}
\usage{
snp_imputeCV2(x, nrounds = 20, max_depth = 3, size = 1000, alpha = 0.1,
  K = 5, ncores = 1, verbose = FALSE, baseline = FALSE)
}
\arguments{
\item{max_depth}{}
}
\description{
Title
}
