% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsc.R
\name{snp_ldsc}
\alias{snp_ldsc}
\title{LD score regression}
\usage{
snp_ldsc(
  ld_score,
  ld_size,
  chi2,
  sample_size,
  blocks = 200,
  chi2_thr1 = 30,
  chi2_thr2 = Inf
)
}
\arguments{
\item{ld_score}{Vector of LD scores.}

\item{ld_size}{Number of variants used to compute \code{ld_score}.}

\item{chi2}{Vector of chi-squared statistics.}

\item{sample_size}{Sample size of GWAS corresponding to chi-squared statistics.
Possibly a vector, or just a single value.}

\item{blocks}{Either a simgle number specifying the number of blocks,
or a vector of integers specifying the block number of each \code{chi2} value.
Default is \code{200}, dividing into 200 blocks of approximately equal size.}

\item{chi2_thr1}{Threshold on \code{chi2} in step 1. Default is \code{30}.}

\item{chi2_thr2}{Threshold on \code{chi2} in step 2. Default is \code{Inf} (none).}
}
\value{
Vector of 4 values:
\enumerate{
\item LDSC regression intercept,
\item SE of this intercept,
\item LDSC regression estimate of heritability,
\item SE of this heritability estimate.
}
}
\description{
Implementation of LDSC regression in R, originally based on
https://helda.helsinki.fi/bitstream/handle/10138/273501/MT_ldsc_hautakangas.pdf
}
