% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-LD.R
\name{snp_ldsplit}
\alias{snp_ldsplit}
\title{Independent LD blocks}
\usage{
snp_ldsplit(corr, thr_r2, grid_param)
}
\arguments{
\item{corr}{Sparse correlation matrix. Usually the output of \code{\link[=snp_cor]{snp_cor()}}.}

\item{thr_r2}{Threshold under which squared correlations are ignored.}

\item{grid_param}{Grid of parameters to consider, a data frame with
\itemize{
\item \verb{$min_size}: Minimum number of variants in each block.
\item \verb{$max_size}: Maximum number of variants in each block.
\item \verb{$lambda}: Penalty coefficient to apply on the size of the blocks.
Using \code{0} would disable this, while using \code{1e-3} could be a good value
to get more smaller blocks.
}}
}
\value{
Input \code{grid_param} as tibble with five extra columns:
\itemize{
\item \verb{$all_cost}: Internal costs computed.
\item \verb{$cost}: The first cost, \code{C[1]}, represents the sum of squared
correlations outside the blocks.
\item \verb{$all_last}: Last index of each block.
\item \verb{$all_size}: Sizes of the blocks.
\item \verb{$n_block}: Number of blocks.
}
}
\description{
Split a correlation matrix in blocks as independent as possible.
This will find the splitting in blocks that minimize the sum of squared
correlation between these blocks (i.e. everything outside these blocks).
}
