% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAX3.R
\name{snp_MAX3}
\alias{snp_MAX3}
\title{MAX3 statistic}
\usage{
snp_MAX3(x, ind.train = rows_along(X.), val = c(0, 0.5, 1))
}
\arguments{
\item{x}{A \link[=bigSNP-class]{bigSNP}.}

\item{ind.train}{An optional vector of the row indices that are used,
for the training part. If not specified, all rows are used.}

\item{val}{Computing \eqn{\smash{\displaystyle\max_{x \in val}}~Z_{CATT}^2(x)}.
Default is \code{c(0, 0.5, 1)} and corresponds to the \emph{MAX3} statistic.
Only \code{c(0, 1)} corresponds to \emph{MAX2}.
And only \code{0.5} corresponds to the Armitage trend test.
Finally, \code{seq(0, 1, length.out = L)} corresponds to \emph{MAXL}.}
}
\value{
A data.frame of \code{S} and \code{pS} for every column, which are MAX3
statistics and associated p-values. \strong{P-values are in fact the minimum of
the p-values of each test separately (so they are biased downward).}
One can use genomic control to rescale these p-values.
}
\description{
Counts the number of 0, 1, 2 by SNP and phenotype (cases/controls)
and then compute the MAX3 statistic.
}
\examples{
set.seed(1)

# constructing a fake genotype big.matrix
a <- big.matrix(10, 12, type = "char", shared = FALSE)
a[] <- sample(c(0, 1, 2), 150, TRUE)
print(a[,])

# constructing a fake incomplete bigSNP with 10 individuals and 15 SNPs
# where the 5 first individuals are cases and the 5 last are controls.
fake <- list()
class(fake) <- "bigSNP"
fake$genotypes <- a
fake$fam$affection <- c(rep(1, 5), rep(-1, 5))

# Get MAX3 statistics
print(snp_MAX3(fake))

}
\references{
Zheng, G., Yang, Y., Zhu, X., & Elston, R. (2012).
Robust Procedures. Analysis Of Genetic Association Studies, 151-206.
\url{http://dx.doi.org/10.1007/978-1-4614-2245-7_6}.
}
