% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-software.R
\name{snp_plinkIBDQC}
\alias{snp_plinkIBDQC}
\title{Identity-by-descent}
\usage{
snp_plinkIBDQC(plink.path, bedfile.in, bedfile.out = NULL, pi.hat = 0.08,
  ncores = 1, pruning.args = c(100, 0.2), extra.options = "")
}
\arguments{
\item{plink.path}{Path to the executable of PLINK 1.9.}

\item{bedfile.in}{Path to the input bedfile.}

\item{bedfile.out}{Path to the output bedfile. Default is created by
appending \code{"_norel"} to \code{prefix.in} (\code{bedfile.in} without extension).}

\item{pi.hat}{PI_HAT value threshold for individuals (the second)
to be excluded. Default is \code{0.08}.}

\item{ncores}{Number of cores to be used. Default is \code{1}. An usually good
value for this parameter is \code{ncores = parallel::detectCores() - 1}.}

\item{pruning.args}{A vector of 2 pruning parameters, respectively
the window size (in variant count) and the pairwise $r^2$ threshold
(the step size is fixed to 1). Default is \code{c(100, 0.2)}.}

\item{extra.options}{Other options to be passed to PLINK as a string. More
options can be found at \url{https://www.cog-genomics.org/plink/1.9/ibd}.}
}
\value{
The path of the new bedfile. If no sample is filter, no new bed/bim/fam
files are created and then the path of the input bedfile is returned.
}
\description{
Quality Control based on Identity-by-descent (IBD) computed by \strong{PLINK 1.9}
using its method-of-moments.
}
\examples{
\dontrun{
# On MY computer, I can do:
plink <- "../plink_linux_x86_64/plink"
bedfile <- "../POPRES_data/POPRES_allchr_QC.bed"
test <- snp_plinkIBDQC(plink, bedfile, ncores = 4)}
}
\references{
Purcell, S., Neale, B., Todd-Brown, K., Thomas, L., Ferreira, M. A.,
Bender, D., ... & Sham, P. C. (2007).
PLINK: a tool set for whole-genome association and population-based linkage
analyses. The American Journal of Human Genetics, 81(3), 559-575.
}
