% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRS.R
\name{snp_PRS}
\alias{snp_PRS}
\title{PRS}
\usage{
snp_PRS(G, betas, ind.test, ind.keep = cols_along(G), lpS = NULL,
  thr.list = 0)
}
\arguments{
\item{G}{A \link[=BM.code.descriptor-class]{BM.code.descriptor}.
Typically the slot \code{genotypes} of a \code{bigSNP}.
You shouldn't have missing values in your data.}

\item{betas}{Numeric vector of weights associated with each SNP.
You may want to see \link{big_univRegLin} or \link{big_univRegLog}.}

\item{ind.test}{The individuals on whom to project the scores.}

\item{ind.keep}{Column (SNP) indices to use (if using clumping, the
output of \link{snp_clumping}). Default doesn't clump.}

\item{lpS}{Numeric vector of \code{-log10(p.value)} associated with \code{betas}.
Default doesn't use thresholding.}

\item{thr.list}{Threshold vector on \code{lpS} at which SNPs are excluded if
they are not significant enough. Default doesn't use thresholding.}
}
\value{
A matrix of scores, where rows correspond to \code{ind.test} and
columns correspond to \code{thr.list}.
}
\description{
Polygenic Risk Scores with possible clumping and thresholding.
}
