% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRS.R
\name{snp_PRS}
\alias{snp_PRS}
\title{PRS}
\usage{
snp_PRS(G, betas, ind.test, ind.keep = cols_along(G), lpS = NULL,
  thr.list = 0)
}
\arguments{
\item{G}{A \link[=BM.code.descriptor-class]{BM.code.descriptor}
(typically \code{<bigSNP>$genotypes}).\cr
\strong{You shouldn't have missing values in your data or SNPs with 0 MAF.}}

\item{betas}{Numeric vector of weights associated with each SNP.
You may want to see \link{big_univLinReg} or \link{big_univLogReg}.}

\item{ind.test}{The individuals on whom to project the scores.}

\item{ind.keep}{Column (SNP) indices to use (if using clumping, the
output of \link{snp_clumping}). Default doesn't clump.}

\item{lpS}{Numeric vector of \code{-log10(p.value)} associated with \code{betas}.
Default doesn't use thresholding.}

\item{thr.list}{Threshold vector on \code{lpS} at which SNPs are excluded if
they are not significant enough. Default doesn't use thresholding.}
}
\value{
A matrix of scores, where rows correspond to \code{ind.test} and
columns correspond to \code{thr.list}.
}
\description{
Polygenic Risk Scores with possible clumping and thresholding.
}
\examples{
test <- snp_attachExtdata()
G <- test$genotypes
n <- nrow(G)
m <- ncol(G)

# get some phenotypes (LTM)
set.seed(5)
X <- attach.BM(G)
M <- 100 # number of causal variants
K <- 0.3 # prevalence
ind.causal <- sample(m, M)
betas <- rnorm(M, sd = sqrt(0.8 / M))
y.g <- drop(scale(X[, ind.causal]) \%*\% betas)
rm(X)
(v <- var(y.g))
y.e <- rnorm(n, sd = sqrt(1 - v))
(y01 <- as.numeric((y.g + y.e) > qnorm(1 - K)))

# PCA -> covariables
obj.svd <- snp_autoSVD(G, infos.chr = test$map$chromosome,
                       infos.pos = test$map$physical.position)

# train and test set
ind.train <- sort(sample(n, 400))
ind.test <- setdiff(rows_along(G), ind.train) # 117

# GWAS
gwas.train <- big_univLogReg(G, y01.train = y01[ind.train],
                             ind.train = ind.train,
                             covar.train = obj.svd$u[ind.train, ])
# clumping
ind.keep <- snp_clumping(G, infos.chr = test$map$chromosome,
                         ind.row = ind.train,
                         S = abs(gwas.train$score))
# -log10(p-values) and thresolding
summary(lpS <- -predict(gwas.train))
thrs <- seq(0, 3.5, by = 0.5)
nb.pred <- sapply(thrs, function(thr) sum(lpS[ind.keep] > thr))

# PRS
prs <- snp_PRS(G, betas = gwas.train$estim,
               ind.test = ind.test,
               ind.keep = ind.keep,
               lpS = lpS,
               thr.list = thrs)

# AUC as a function of the number of predictors
aucs <- apply(prs, 2, AUC, target = y01[ind.test])
library(ggplot2)
bigstatsr:::MY_THEME(qplot(nb.pred, aucs)) +
  geom_line() +
  scale_x_log10(breaks = nb.pred) +
  labs(x = "Number of predictors", y = "AUC")
}
