% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-sumstats.R
\name{snp_qc_sumstats}
\alias{snp_qc_sumstats}
\title{Quality Control of summary statistics}
\usage{
snp_qc_sumstats(
  z_sumstats,
  ld,
  thr_highld = 0.05,
  num_highld = 20,
  prop_eig = 0.4,
  max_run = 0.1 * length(z_sumstats),
  print_info = FALSE,
  ncores = 1
)
}
\arguments{
\item{z_sumstats}{z_sumstats from GWAS summary statistics}

\item{ld}{LD matrix}

\item{thr_highld}{Threshold for highly correlated variants, \code{0.05} by default}

\item{num_highld}{Minimum number of highly correlated variants, \code{20} by default}

\item{prop_eig}{Proportion of eigenvalues to attain, \code{0.4} by default}

\item{max_run}{Maximum round of iterations to run.}

\item{print_info}{print_info or not, \code{FALSE} by default}

\item{ncores}{Number of cores to run in parallel}
}
\value{
A dataframe of original z-score, imputed z-score, chisq statistics, p-value, and whether to be removed or not
}
\description{
Quality Control of summary statistics
}
\examples{


obj_bigsnp <- snp_attachExtdata()
G <- obj_bigsnp$genotypes
Z <- big_univLogReg(G, obj_bigsnp$fam$affection - 1, ind.col = 1:500)$score
ind <- which.max(abs(Z))
Z[ind] <- -Z[ind]
ld <- as_SFBM(snp_cor(G, ind.col = 1:500), compact = TRUE)
res <- snp_qc_sumstats(z_sumstats = Z, ld = ld)
res$p_val[ind]
res$remove[p_val]

}
