% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-sumstats.R
\name{snp_qc_sumstats}
\alias{snp_qc_sumstats}
\title{Quality Control of summary statistics}
\usage{
snp_qc_sumstats(
  z_scores,
  ld,
  thr_highld = 0.05,
  num_highld = 20,
  prop_eig = 0.4,
  max_run = 0.1 * length(z_scores),
  ncores = 1
)
}
\arguments{
\item{z_scores}{z_scores from GWAS summary statistics}

\item{ld}{LD matrix}

\item{thr_highld}{Threshold for highly correlated variants, \code{0.05} by default}

\item{num_highld}{Minimum number of highly correlated variants, \code{20} by default}

\item{prop_eig}{Proportion of eigenvalues to attain, \code{0.4} by default}

\item{max_run}{Maximum round of iterations to run.}

\item{ncores}{Number of cores to run in parallel}
}
\value{
A dataframe of original z-score, imputed z-score, chisq statistics, p-value, and whether to be removed or not
}
\description{
Quality Control of summary statistics
}
