% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readplink.R
\name{snp_readBed}
\alias{snp_readBed}
\title{Read PLINK files into a "bigSNP"}
\usage{
snp_readBed(bedfile, block.size = 1000, backingfile,
  backingpath = "backingfiles")
}
\arguments{
\item{bedfile}{Path to file with extension ".bed" to read.
You need the corresponding ".bim" and ".fam" in the same directory.}

\item{backingfile}{The root name for the backing file(s) for the cache of
the \link[=bigSNP-class]{bigSNP} object.}

\item{backingpath}{The path to the directory containing the backing files.
Default is \code{"backingfiles"}.}
}
\value{
The path to one of the backing files which are created
by this function: \code{<backingfile>}.bk, \code{<backingfile>}.desc and
\code{<backingfile>}.rds in directory \code{<backingpath>}.\cr
\strong{You shouldn't read from PLINK files more than once.} Instead, use
\link{snp_attach} to load the "bigSNP" object in any R session from backing files.
}
\description{
Function to read bed/bim/fam files into a \link[=bigSNP-class]{bigSNP}.
}
\details{
For more information on these formats, please visit
\href{http://pngu.mgh.harvard.edu/~purcell/plink/data.shtml#bed}{PLINK webpage}.
For other formats, please use PLINK to convert them in bedfiles,
which require minimal space to store and are faster to read.
}
\note{
The implementation was originally inspired from the code of
\href{https://github.com/andrewparkermorgan/argyle}{package argyle}.
I did many optimizations. Moreover, online reading
into a \code{big.matrix} makes it memory-efficient.
}
\examples{
bedfile <- system.file("extdata", "example.bed", package = "bigsnpr")
print(bedfile)

# Reading the bedfile and storing the data in directory "backingfiles"
PATH <- "backingfiles/test_doc.bk"
if (!file.exists(PATH))
  PATH <- snp_readBed(bedfile, backingfile = "test_doc")

# Loading the data from backing files
test <- snp_attach(PATH)

str(test)
print(dim(test$genotypes))
print(test$genotypes[1:8, 1:8])
}
