% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readbgen.R
\name{snp_readBGEN}
\alias{snp_readBGEN}
\title{Read BGEN files into a "bigSNP"}
\usage{
snp_readBGEN(bgenfiles, backingfile, snp_id)
}
\arguments{
\item{bgenfiles}{Character vector of paths to files with extension ".bgen".}

\item{backingfile}{The path (without extension) for the backing files
for the cache of the \link[=bigSNP-class]{bigSNP} object.}

\item{snp_id}{Character vector of SNP IDs (not rsid) to read
(e.g. "1:88169_C_T" in the UK Biobank).}
}
\value{
The path to the RDS file that stores the \code{bigSNP} object.
Note that this function creates one other file which stores the values of
the Filebacked Big Matrix.\cr
\strong{You shouldn't read from BGEN files more than once.} Instead, use
\link{snp_attach} to load the "bigSNP" object in any R session from backing files.
}
\description{
Function to read the UK Biobank BGEN files into a \link[=bigSNP-class]{bigSNP}.
}
\details{
For more information on this format, please visit
\href{https://bitbucket.org/gavinband/bgen/}{BGEN webpage}.

This function is designed to read UK Biobank imputation files. This assumes
that variants have been compressed with zlib, that there are only 2 possible
alleles, and that each probability is stored on 8 bits.
}
