% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveModifs.R
\name{snp_saveModifs}
\alias{snp_saveModifs}
\title{Save modifications for a "bigSNP"}
\usage{
snp_saveModifs(x)
}
\arguments{
\item{x}{A \link[=bigSNP-class]{bigSNP}.}
}
\value{
The (saved) \code{bigSNP}.
}
\description{
Save a \code{bigSNP} after having made some modifications to it.
As \code{bigSNP} is an S3 class, you can add any slot you want
to an object of this class, then use \code{snp_saveModifs} to
save these modifications in the corresponding .rds backing file.
}
\examples{
#reading example
test <- snp_readExample()

# I can add whatever I want to a S3 class
test$map$`p-values` <- runif(nrow(test$map))
str(test$map)

# reading again
test <- snp_readExample()
str(test$map) # new slot wasn't saved

# save it
test$map$`p-values` <- runif(nrow(test$map))
test <- snp_saveModifs(test)

# reading again
test <- snp_readExample()
str(test$map) # it is saved now

}

