% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readplink.R
\name{snp_writeBed}
\alias{snp_writeBed}
\title{Write PLINK files from a "bigSNP"}
\usage{
snp_writeBed(x, bedfile, ind.row = rows_along(X), ind.col = cols_along(X))
}
\arguments{
\item{x}{A \link[=bigSNP-class]{bigSNP}.}

\item{bedfile}{Path to file with extension ".bed" to create.}

\item{ind.row}{An optional vector of the row indices (individuals) that
are used. If not specified, all rows are used.\cr
\strong{Don't use negative indices.}}

\item{ind.col}{An optional vector of the column indices (SNPs) that are used.
If not specified, all columns are used.\cr
\strong{Don't use negative indices.}}
}
\value{
The input \code{bedfile} path.
}
\description{
Function to write bed/bim/fam files from a \link[=bigSNP-class]{bigSNP}.
This will use the slot \code{code} \strong{rounded} to write 0s, 1s, 2s or NAs.
}
\examples{
N <- 17
M <- 911

fake <- snp_fake(N, M)
X <- attach.BM(fake$genotypes)
X[] <- sample(as.raw(0:3), length(X), TRUE)
rm(X)

# write the object as a bed/bim/fam object
tmp <- tempfile(fileext = ".bed")
bed <- snp_writeBed(fake, tmp)
# read this new file for the first time
rds <- snp_readBed(bed, backingfile = "test_write")
# attach object in R session
fake2 <- snp_attach(rds)

# same content
all.equal(attach.BM(fake$genotypes)[,],
          attach.BM(fake2$genotypes)[,])
all.equal(fake$fam, fake2$fam)
all.equal(fake$map, fake2$map)

# two different files
fake$savedIn
fake2$savedIn



# cleaning
bk <- sub("\\\\.rds$", ".bk", rds)
file.remove(rds, bk)
}
